% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_cell_type_proportions}
\alias{add_cell_type_proportions}
\title{Add cell type proportions from cibersort}
\usage{
add_cell_type_proportions(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  reference = X_cibersort,
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{reference}{A data frame. The transcript/cell_type data frame of integer transcript abundance}

\item{...}{Further parameters passed to the function Cibersort}
}
\value{
A tibble including additional columns
}
\description{
Add cell type proportions from cibersort
}
