% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_normalised_counts_bulk.calcNormFactor}
\alias{add_normalised_counts_bulk.calcNormFactor}
\title{Calculate the norm factor with calcNormFactor from limma}
\usage{
add_normalised_counts_bulk.calcNormFactor(df, reference = NULL,
  cpm_threshold = 0.5, prop = 3/4, sample_column = sample,
  transcript_column = transcript, counts_column = count)
}
\arguments{
\item{df}{A tibble}

\item{reference}{A reference matrix, not sure if used anymore}

\item{cpm_threshold}{A real positive number}

\item{prop}{A number between 0 and 1}
}
\value{
A list including the filtered data frame and the normalization factors
}
\description{
Calculate the norm factor with calcNormFactor from limma
}
