% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_normalised_counts_bulk.get_cpm}
\alias{add_normalised_counts_bulk.get_cpm}
\title{Get count per million for TMM normalisation.}
\usage{
add_normalised_counts_bulk.get_cpm(df, sample_column = sample,
  transcript_column = transcript, counts_column = count,
  cpm_threshold = 0.5)
}
\arguments{
\item{df}{A tibble}

\item{cpm_threshold}{A real positive number}
}
\value{
A tibble with an additional column
}
\description{
Get count per million for TMM normalisation.
}
