% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_normalised_counts_bulk.get_low_expressed}
\alias{add_normalised_counts_bulk.get_low_expressed}
\title{Drop lowly tanscribed genes for TMM normalization}
\usage{
add_normalised_counts_bulk.get_low_expressed(input.df,
  sample_column = sample, transcript_column = transcript,
  counts_column = count, cpm_threshold = 0.5, prop = 3/4)
}
\arguments{
\item{input.df}{A tibble}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}

\item{cpm_threshold}{A real positive number}

\item{prop}{A number between 0 and 1}
}
\value{
A tibble filtered
}
\description{
Drop lowly tanscribed genes for TMM normalization
}
