% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_normalised_counts_bulk}
\alias{add_normalised_counts_bulk}
\title{Add a tibble with normalised counts using TMM}
\usage{
add_normalised_counts_bulk(input.df, sample_column = NULL,
  transcript_column = NULL, counts_column = NULL,
  cpm_threshold = 0.5, prop = 3/4, method = "TMM",
  reference_selection_function = median)
}
\arguments{
\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}

\item{cpm_threshold}{A real positive number}

\item{prop}{A number between 0 and 1}

\item{method}{A character string}

\item{df}{A tibble}

\item{reference}{A reference matrix, not sure if used anymore}
}
\value{
A tibble including additional columns
}
\description{
Add a tibble with normalised counts using TMM
}
