% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_reduced_dimensions_TSNE_bulk}
\alias{add_reduced_dimensions_TSNE_bulk}
\title{Add principal component information to a tibble using tSNE}
\usage{
add_reduced_dimensions_TSNE_bulk(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  .dims = 2,
  top = 500,
  of_samples = T,
  log_transform = T,
  scale = T,
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A column symbol. The column that is used to calculate distance (i.e., normally samples)}

\item{.feature}{A column symbol. The column that is represents entities to cluster (i.e., normally genes)}

\item{.abundance}{A column symbol with the value the clustering is based on (e.g., `count`)}

\item{.dims}{A integer vector corresponding to principal components of interest (e.g., 1:6)}

\item{top}{An integer. How many top genes to select}

\item{of_samples}{A boolean}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{scale}{A boolean}

\item{...}{Further parameters passed to the function tSNE}
}
\value{
A tibble with additional columns
}
\description{
Add principal component information to a tibble using tSNE
}
