% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_scaled_counts_bulk.calcNormFactor}
\alias{add_scaled_counts_bulk.calcNormFactor}
\title{Calculate the norm factor with calcNormFactor from limma}
\usage{
add_scaled_counts_bulk.calcNormFactor(
  .data,
  reference = NULL,
  .sample = sample,
  .transcript = transcript,
  .abundance = count,
  method,
  .library_size = NULL
)
}
\arguments{
\item{.data}{A tibble}

\item{reference}{A reference matrix, not sure if used anymore}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{method}{A string character. The scaling method passed to the backend function (i.e., edgeR::calcNormFactors; "TMM","TMMwsp","RLE","upperquartile")}
}
\value{
A list including the filtered data frame and the normalization factors
}
\description{
Calculate the norm factor with calcNormFactor from limma
}
\keyword{internal}
