% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_scaled_counts_bulk}
\alias{add_scaled_counts_bulk}
\title{Add a tibble with scaled counts using TMM}
\usage{
add_scaled_counts_bulk(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  cpm_threshold = 0.5,
  prop_threshold = 3/4,
  method = "TMM",
  reference_selection_function = median
)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{cpm_threshold}{A real positive number. Minimum counts per million required for a selected proportion of samples}

\item{prop_threshold}{A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for normalisation procedure.}

\item{method}{A character string}

\item{reference_selection_function}{A function between median, mean and max}
}
\value{
A tibble including additional columns
}
\description{
Add a tibble with scaled counts using TMM
}
