% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_symbol_from_ensembl}
\alias{add_symbol_from_ensembl}
\title{Add transcript column from ensembl gene id}
\usage{
add_symbol_from_ensembl(.data, .ensembl)
}
\arguments{
\item{.data}{A tibble}

\item{.ensembl}{A column symbol. The column that is represents ensembl gene id}
}
\value{
A tibble with added annotation
}
\description{
Add transcript column from ensembl gene id
}
