% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{adjust_abundance,spec_tbl_df-method}
\alias{adjust_abundance,spec_tbl_df-method}
\title{adjust_abundance}
\usage{
\S4method{adjust_abundance}{spec_tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  log_transform = TRUE,
  action = "add",
  ...
)
}
\description{
adjust_abundance
}
