% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{adjust_abundance}
\alias{adjust_abundance}
\title{Adjust transcript abundance for unwanted variation}
\usage{
adjust_abundance(.data, .formula, .sample = NULL, .transcript = NULL,
  .abundance = NULL, log_transform = T, action = "add", ...)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model where the first covariate is the factor of interest and the second covariate is the unwanted variation (of the kind ~ factor_of_intrest + batch)}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{action}{A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).}

\item{...}{Further parameters passed to the function sva::ComBat}
}
\value{
A `tbl` with additional columns for the adjusted counts as `<COUNT COLUMN> adjusted`
}
\description{
adjust_abundance() takes as imput a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with an edditional adjusted abundance column..
}
\details{
This function adjusts the abundance for (known) unwanted variation. At the moment just an unwanted covariated is allowed at a time.
}
\examples{
\donttest{



adjust_abundance(
    ~ factor_of_interest + batch,
    sample,
    transcript,
    `count`
)

}

}
