% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{adjust_counts}
\alias{adjust_counts}
\title{Get adjusted count for unwanted variation}
\usage{
adjust_counts(input.df, formula, sample_column = NULL,
  transcript_column = NULL, counts_column = NULL, log_transform = T,
  action = "add", ...)
}
\arguments{
\item{input.df}{A tibble}

\item{formula}{a formula with no response variable, of the kind ~ factor_of_intrest + batch}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}

\item{log_transform}{A boolean. For correlation based calculation. Whether the value should be log-transformed (e.g., TRUE for RNA sequencing data).}

\item{action}{A character string. Whether to join the new information to the input tibble (add), or just get the non-redundant tibble with the new information (get).}

\item{...}{Further parameters passed to the function sva::ComBat}
}
\value{
A tibble with edgeR results
}
\description{
Get adjusted count for unwanted variation
}
