% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{aggregate_duplicated_transcripts_bulk}
\alias{aggregate_duplicated_transcripts_bulk}
\title{Aggregates multiple counts from the same samples (e.g., from isoforms)
This function aggregates counts over samples, concatenates other character columns, and averages other numeric columns}
\usage{
aggregate_duplicated_transcripts_bulk(input.df,
  aggregation_function = sum, sample_column = NULL,
  transcript_column = NULL, counts_column = NULL, keep_integer = T)
}
\arguments{
\item{input.df}{A tibble}

\item{aggregation_function}{A function for counts aggregation (e.g., sum)}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}
}
\value{
A tibble with aggregated genes and annotation
}
\description{
Aggregates multiple counts from the same samples (e.g., from isoforms)
This function aggregates counts over samples, concatenates other character columns, and averages other numeric columns
}
