% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aggregate_duplicated_transcripts_DT}
\alias{aggregate_duplicated_transcripts_DT}
\title{Aggregates multiple counts from the same samples (e.g., from isoforms)
This function aggregates counts over samples, concatenates other character columns, and averages other numeric columns}
\usage{
aggregate_duplicated_transcripts_DT(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  aggregation_function = sum,
  keep_integer = TRUE
)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{aggregation_function}{A function for counts aggregation (e.g., sum)}

\item{keep_integer}{A boolean}
}
\value{
A tibble with aggregated genes and annotation
}
\description{
Aggregates multiple counts from the same samples (e.g., from isoforms)
This function aggregates counts over samples, concatenates other character columns, and averages other numeric columns
}
\keyword{internal}
