% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{aggregate_duplicates,RangedSummarizedExperiment-method}
\alias{aggregate_duplicates,RangedSummarizedExperiment-method}
\title{aggregate_duplicates}
\usage{
\S4method{aggregate_duplicates}{RangedSummarizedExperiment}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  aggregation_function = sum,
  keep_integer = TRUE
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{aggregation_function}{A function for counts aggregation (e.g., sum,  median, or mean)}

\item{keep_integer}{A boolean. Whether to force the aggregated counts to integer}
}
\value{
A `tidyBulk` object
}
\description{
aggregate_duplicates
}
