% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{aggregate_duplicates,tbl_df-method}
\alias{aggregate_duplicates,tbl_df-method}
\title{aggregate_duplicates}
\usage{
\S4method{aggregate_duplicates}{tbl_df}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  aggregation_function = sum,
  keep_integer = TRUE
)
}
\description{
aggregate_duplicates
}
