% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{analise_gene_enrichment}
\alias{analise_gene_enrichment}
\title{Analise gene enrichment with EGSEA}
\usage{
analise_gene_enrichment(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  species,
  cores = 10
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model}

\item{.sample}{The name of the sample column}

\item{.entrez}{The ENTREZ doce of the transcripts/genes}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.contrasts}{= NULL,}

\item{species}{A character. For example, human or mouse}

\item{cores}{An integer. The number of cores available}
}
\value{
A `tbl` object
}
\description{
analise_gene_enrichment() takes as imput a `tbl` formatted as | <SAMPLE> | <ENSEMBL_ID> | <COUNT> | <...> | and returns a `tbl` with the the additional transcript symbol column
}
\details{
This wrapper execute gene enrichment analyses of the dataset
}
\examples{
\donttest{



	analise_gene_enrichment(
		input_df,
		~ condition,
		.sample = a,
		.entrez = d,
		.abundance = c,
		species="human"
	)

}

}
