% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{annotate_cell_type}
\alias{annotate_cell_type}
\title{Get cell type proportions from cibersort}
\usage{
annotate_cell_type(input.df, sample_column = NULL,
  transcript_column = NULL, counts_column = NULL, action = "add",
  ...)
}
\arguments{
\item{input.df}{A tibble}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}

\item{action}{A character string. Whether to join the new information to the input tibble (add), or just get the non-redundant tibble with the new information (get).}

\item{...}{Further parameters passed to the function Cibersort}
}
\value{
A tibble including additional columns
}
\description{
Get cell type proportions from cibersort
}
