% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{annotate_clusters}
\alias{annotate_clusters}
\title{Get clusters of elements (e.g., samples or transcripts)}
\usage{
annotate_clusters(input.df, value_column, number_of_clusters,
  elements_column = NULL, feature_column = NULL, of_samples = T,
  log_transform = T, action = "add")
}
\arguments{
\item{input.df}{A tibble}

\item{value_column}{A character string. The column that contains the numeric value (i.e., normally read counts)}

\item{number_of_clusters}{A integer indicating how many clusters we are seeking}

\item{elements_column}{A character string. The column that is used to calculate distance (i.e., normally genes)}

\item{feature_column}{A character string. The column that is represents entities to cluster (i.e., normally samples)}

\item{of_samples}{if the input is tt object, this will indicate whether the element column will be sample or transcript column}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{action}{A character string. Whether to join the new information to the input tibble (add), or just get the non-redundant tibble with the new information (get).}
}
\value{
A tibble with additional columns
}
\description{
Get clusters of elements (e.g., samples or transcripts)
}
