% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{annotate_differential_transcription}
\alias{annotate_differential_transcription}
\title{Add differential transcription information to a tibble using edgeR. At the moment only one covariate is accepted}
\usage{
annotate_differential_transcription(input.df, formula,
  sample_column = NULL, transcript_column = NULL,
  counts_column = NULL, significance_threshold = 0.05,
  action = "add")
}
\arguments{
\item{input.df}{A tibble}

\item{formula}{a formula with no response variable, referring only to numeric variables}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}

\item{significance_threshold}{A real between 0 and 1.}

\item{action}{A character string. Whether to join the new information to the input tibble (add), or just get the non-redundant tibble with the new information (get).}
}
\value{
A tibble with differential_transcript_abundance results
}
\description{
Add differential transcription information to a tibble using edgeR. At the moment only one covariate is accepted
}
