% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{annotate_symbol,spec_tbl_df-method}
\alias{annotate_symbol,spec_tbl_df-method}
\title{annotate_symbol}
\usage{
\S4method{annotate_symbol}{spec_tbl_df}(.data, .ensembl, action = "add")
}
\description{
annotate_symbol
}
