% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{annotate_symbol}
\alias{annotate_symbol}
\title{Add transcript column from ensembl gene id}
\usage{
annotate_symbol(input.df, ensembl_transcript_column, action = "add")
}
\arguments{
\item{input.df}{A tibble}

\item{ensembl_transcript_column}{A character string. The column that is represents ensembl gene id}

\item{action}{A character string. Whether to join the new information to the input tibble (add), or just get the non-redundant tibble with the new information (get).}
}
\value{
A tibble with added annotation
}
\description{
Add transcript column from ensembl gene id
}
