% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{bind_rows}
\alias{bind_rows}
\title{Join datasets}
\usage{
bind_rows(...)
}
\arguments{
\item{...}{Data frames to combine}
}
\value{
A tt object
}
\description{
Join datasets
}
\examples{

bind_rows(ttBulk::counts_mini, ttBulk::counts_mini)


}
