% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast_tcga_mini}
\alias{breast_tcga_mini}
\title{Data set}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 125500 rows and 5 columns.}
\usage{
breast_tcga_mini
}
\description{
Data set
}
\keyword{datasets}
