% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast_tcga_mini_SE}
\alias{breast_tcga_mini_SE}
\title{Needed for vignette breast_tcga_mini_SE}
\format{
An object of class \code{SummarizedExperiment} with 500 rows and 251 columns.
}
\usage{
breast_tcga_mini_SE
}
\description{
Needed for vignette breast_tcga_mini_SE
}
\keyword{datasets}
