% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cluster_elements,ttBulk-method}
\alias{cluster_elements,ttBulk-method}
\title{cluster_elements}
\usage{
\S4method{cluster_elements}{ttBulk}(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  method,
  of_samples = TRUE,
  log_transform = TRUE,
  action = "add",
  ...
)
}
\description{
cluster_elements
}
