% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_ensembl}
\alias{counts_ensembl}
\title{Counts with ensembl annotation}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 119 rows and 6 columns.}
\usage{
counts_ensembl
}
\description{
Counts with ensembl annotation
}
\keyword{datasets}
