% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_mini}
\alias{counts_mini}
\title{Example data set reduced}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2635 rows and 6 columns.
}
\usage{
counts_mini
}
\description{
Example data set reduced
}
\keyword{datasets}
