% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_SE}
\alias{counts_SE}
\title{Needed for vignette counts_SE}
\format{
An object of class \code{SummarizedExperiment} with 8513 rows and 48 columns.
}
\usage{
counts_SE
}
\description{
Needed for vignette counts_SE
}
\keyword{datasets}
