% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{create_ttBulk}
\alias{create_ttBulk}
\title{Create tt object from tibble}
\usage{
create_ttBulk(input.df, sample_column, transcript_column, counts_column)
}
\arguments{
\item{input.df}{A tibble}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}
}
\value{
A tibble with an additional column
}
\description{
Create tt object from tibble
}
