% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{deconvolve_cellularity,spec_tbl_df-method}
\alias{deconvolve_cellularity,spec_tbl_df-method}
\title{deconvolve_cellularity}
\usage{
\S4method{deconvolve_cellularity}{spec_tbl_df}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  reference = X_cibersort,
  method = "cibersort",
  action = "add",
  ...
)
}
\description{
deconvolve_cellularity
}
