% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_SE.R
\docType{methods}
\name{describe_transcript}
\alias{describe_transcript}
\alias{describe_transcript,spec_tbl_df-method}
\alias{describe_transcript,tbl_df-method}
\alias{describe_transcript,tidybulk-method}
\alias{.describe_transcript_SE}
\alias{describe_transcript,SummarizedExperiment-method}
\alias{describe_transcript,RangedSummarizedExperiment-method}
\title{Get DESCRIPTION from gene SYMBOL for Human and Mouse}
\usage{
describe_transcript(.data, .transcript = NULL)

\S4method{describe_transcript}{spec_tbl_df}(.data, .transcript = NULL)

\S4method{describe_transcript}{tbl_df}(.data, .transcript = NULL)

\S4method{describe_transcript}{tidybulk}(.data, .transcript = NULL)

.describe_transcript_SE(.data, .transcript = NULL)

\S4method{describe_transcript}{SummarizedExperiment}(.data, .transcript = NULL)

\S4method{describe_transcript}{RangedSummarizedExperiment}(.data, .transcript = NULL)
}
\arguments{
\item{.data}{A tt or tbl object.}

\item{.transcript}{A character. The name of the gene symbol column.}
}
\value{
A tbl

A `tbl` object including additional columns for transcript symbol

A `tbl` object including additional columns for transcript symbol

A `tbl` object including additional columns for transcript symbol

A `SummarizedExperiment` object

A `tbl` object including additional columns for transcript symbol

A `tbl` object including additional columns for transcript symbol
}
\description{
Get DESCRIPTION from gene SYMBOL for Human and Mouse

describe_transcript

describe_transcript

describe_transcript

describe_transcript

describe_transcript

describe_transcript
}
\examples{

describe_transcript(tidybulk::counts_mini, .transcript = transcript)

}
