% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{describe_transcript}
\alias{describe_transcript}
\title{Get DESCRIPTION id from gene SYMBOL}
\usage{
describe_transcript(.data, .transcript = NULL)
}
\arguments{
\item{.data}{A tt or tbl object.}

\item{.transcript}{A character. The name of the gene symbol column.}
}
\value{
A tbl
}
\description{
Get DESCRIPTION id from gene SYMBOL
}
\examples{

describe_transcript(tidybulk::counts_mini, .transcript = transcript)

}
