% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{distinct}
\alias{distinct}
\title{distinct}
\arguments{
\item{.data}{A tbl. (See dplyr)}

\item{...}{Data frames to combine (See dplyr)}

\item{.keep_all}{If TRUE, keep all variables in .data. If a combination of ... is not distinct, this keeps the first row of values. (See dplyr)}
}
\value{
A tt object
}
\description{
distinct
}
\examples{

tidybulk::se_mini \%>\% tidybulk() \%>\% distinct()


}
