% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{drop_attr}
\alias{drop_attr}
\title{Drop attribute to abject}
\usage{
drop_attr(var, name)
}
\arguments{
\item{var}{A tibble}

\item{name}{A character name of the attribute}
}
\value{
A tibble with an additional attribute
}
\description{
Drop attribute to abject
}
\keyword{internal}
