% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{drop_redundant}
\alias{drop_redundant}
\title{Drop redundant elements (e.g., samples) for which feature (e.g., genes) aboundances are correlated}
\usage{
drop_redundant(input.df, method, elements_column = NULL,
  of_samples = T, value_column, feature_column = NULL,
  correlation_threshold = 0.9, log_transform = F, Dim_a_column,
  Dim_b_column)
}
\arguments{
\item{input.df}{A tibble}

\item{method}{A string character. Method used for calculating the redundancy (e.g., correlation, reduced_dimensions proximity)}

\item{elements_column}{A character string. The column that is used to calculate distance (i.e., normally samples)}

\item{of_samples}{if the input is tt object, this will indicate whether the element column will be sample or transcript column}

\item{value_column}{A character string. For correlation based calculation. The column that contains the numeric value (i.e., normally read counts)}

\item{feature_column}{A character string. For correlation based calculation. The column that is represents entities to cluster (i.e., normally genes)}

\item{correlation_threshold}{A real number between 0 and 1. For correlation based calculation.}

\item{log_transform}{A boolean. For correlation based calculation. Whether the value should be log-transformed (e.g., TRUE for RNA sequencing data).}

\item{Dim_a_column}{A character string. For reduced_dimension based calculation. The column of one principal component}

\item{Dim_b_column}{A character string. For reduced_dimension based calculation. The column of another principal component}
}
\value{
A tibble with redundant elemens removed
}
\description{
Drop redundant elements (e.g., samples) for which feature (e.g., genes) aboundances are correlated
}
