% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{drop_redundant_elements_though_reduced_dimensions}
\alias{drop_redundant_elements_though_reduced_dimensions}
\title{Identifies the closest pairs in a MDS contaxt and return one of them}
\usage{
drop_redundant_elements_though_reduced_dimensions(input.df, Dim_a_column,
  Dim_b_column, elements_column = NULL, of_samples = T)
}
\arguments{
\item{input.df}{A tibble}

\item{Dim_a_column}{A character string. The column of one principal component}

\item{Dim_b_column}{A character string. The column of another principal component}

\item{elements_column}{A character string. The column that is represents entities to cluster (i.e., normally samples)}

\item{counts_column}{A character string. The column that contains the numeric value (i.e., normally counts)}
}
\value{
A tibble with pairs dropped
}
\description{
Identifies the closest pairs in a MDS contaxt and return one of them
}
