% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensembl_symbol_mapping}
\alias{ensembl_symbol_mapping}
\title{Data set}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 291249 rows and 3 columns.
}
\usage{
ensembl_symbol_mapping
}
\description{
Data set
}
\keyword{datasets}
