% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ensembl_to_symbol,tbl_df-method}
\alias{ensembl_to_symbol,tbl_df-method}
\title{ensembl_to_symbol}
\usage{
\S4method{ensembl_to_symbol}{tbl_df}(.data, .ensembl, action = "add")
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <ENSEMBL_ID> | <COUNT> | <...> |}

\item{.ensembl}{A character string. The column that is represents ensembl gene id}

\item{action}{A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).}
}
\value{
A `tbl` object including additional columns for transcript symbol
}
\description{
ensembl_to_symbol
}
