% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{error_if_counts_is_na}
\alias{error_if_counts_is_na}
\title{Check whether there are NA counts}
\usage{
error_if_counts_is_na(input.df, counts_column)
}
\arguments{
\item{input.df}{A tibble of read counts}

\item{counts_column}{A character name of the read count column}
}
\description{
Check whether there are NA counts
}
