% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{error_if_duplicated_genes}
\alias{error_if_duplicated_genes}
\title{Check whether there are duplicated genes/transcripts}
\usage{
error_if_duplicated_genes(input.df, sample_column = sample,
  transcript_column = transcript, counts_column = `read count`)
}
\arguments{
\item{input.df}{A tibble of read counts}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the read count column}
}
\description{
Check whether there are duplicated genes/transcripts
}
