% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{error_if_log_transformed}
\alias{error_if_log_transformed}
\title{Check whether a numeric vector has been log transformed}
\usage{
error_if_log_transformed(x, counts_column)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Check whether a numeric vector has been log transformed
}
