% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{error_if_wrong_input}
\alias{error_if_wrong_input}
\title{Check whether there are NA counts}
\usage{
error_if_wrong_input(input.df, list_input, expected_type)
}
\arguments{
\item{input.df}{A tibble of read counts}

\item{counts_column}{A character name of the read count column}
}
\description{
Check whether there are NA counts
}
