% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{fill_NA_using_formula}
\alias{fill_NA_using_formula}
\title{This function is needed for DE in case the matrix is not rectangular, but includes NA}
\usage{
fill_NA_using_formula(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .abundance_scaled = NULL
)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{a formula with no response variable, of the kind ~ factor_of_intrest + batch}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.abundance_scaled}{The name of the transcript/gene scaled abundance column}
}
\value{
A tibble with adjusted counts
}
\description{
This function is needed for DE in case the matrix is not rectangular, but includes NA
}
