% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fill_NA_with_row_median}
\alias{fill_NA_with_row_median}
\title{This function is needed for DE in case the matrix is not rectangular, but includes NA}
\usage{
fill_NA_with_row_median(.matrix)
}
\arguments{
\item{.matrix}{A matrix}
}
\value{
A matrix
}
\description{
This function is needed for DE in case the matrix is not rectangular, but includes NA
}
