% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{filter_abundant,spec_tbl_df-method}
\alias{filter_abundant,spec_tbl_df-method}
\title{filter_abundant}
\usage{
\S4method{filter_abundant}{spec_tbl_df}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  cpm_threshold = 0.5,
  prop = 3/4
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{cpm_threshold}{A real positive number}

\item{prop}{A number between 0 and 1}
}
\description{
filter_abundant
}
