% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{filter_variable,RangedSummarizedExperiment-method}
\alias{filter_variable,RangedSummarizedExperiment-method}
\title{filter_variable}
\usage{
\S4method{filter_variable}{RangedSummarizedExperiment}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  top = 500,
  log_transform = TRUE
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{top}{Integer. Number of top transcript to consider}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}
}
\value{
A `ttBulk` object
}
\description{
filter_variable
}
