% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{filter_variable}
\alias{filter_variable}
\title{Identify variable transcripts}
\usage{
filter_variable(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  top = 500,
  log_transform = TRUE
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{top}{Integer. Number of top transcript to consider}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}
}
\value{
A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).
}
\description{
filter_variable() takes as imput a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.
}
\details{
At the moment this function uses edgeR only, but other inference algorithms will be added in the near future.
}
\examples{



	filter_variable(
	ttBulk::counts_mini,
	    sample,
	    transcript,
	    `count`,
	    top = 500
	)


}
