% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_variable_transcripts}
\alias{filter_variable_transcripts}
\title{Identify variable genes for dimensionality reduction}
\usage{
filter_variable_transcripts(.data, .sample = NULL, .transcript = NULL,
  .abundance = NULL, top = 3000)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{A character name of the sample column}

\item{.transcript}{A character name of the transcript/gene column}

\item{.abundance}{A character name of the read count column}

\item{top}{An integer. How many top genes to select}
}
\value{
A tibble filtered genes
}
\description{
Identify variable genes for dimensionality reduction
}
