% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SE.R
\name{get_adjusted_counts_for_unwanted_variation_bulk_SE}
\alias{get_adjusted_counts_for_unwanted_variation_bulk_SE}
\title{Get adjusted count for some batch effect}
\usage{
get_adjusted_counts_for_unwanted_variation_bulk_SE(
  .data,
  .formula,
  sample_annotation,
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{a formula with no response variable, of the kind ~ factor_of_interest + batch}

\item{...}{Further parameters passed to the function sva::ComBat}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}
}
\value{
A tibble with adjusted counts
}
\description{
Get adjusted count for some batch effect
}
\keyword{internal}
