% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_clusters_kmeans_bulk}
\alias{get_clusters_kmeans_bulk}
\title{Get K-mean clusters to a tibble}
\usage{
get_clusters_kmeans_bulk(.data, .element = NULL, .feature = NULL,
  .abundance = NULL, of_samples = T, log_transform = T, ...)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A column symbol. The column that is used to calculate distance (i.e., normally genes)}

\item{.feature}{A column symbol. The column that is represents entities to cluster (i.e., normally samples)}

\item{.abundance}{A column symbol with the value the clustering is based on (e.g., `count`)}

\item{of_samples}{A boolean}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{...}{Further parameters passed to the function kmeans}
}
\value{
A tibble with additional columns
}
\description{
Get K-mean clusters to a tibble
}
