% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_clusters_SNN_bulk}
\alias{get_clusters_SNN_bulk}
\title{Get SNN shared nearest neighbour clusters to a tibble}
\usage{
get_clusters_SNN_bulk(.data, .element = NULL, .feature = NULL, .value,
  of_samples = T, log_transform = T, ...)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A column symbol. The column that is used to calculate distance (i.e., normally genes)}

\item{.feature}{A column symbol. The column that is represents entities to cluster (i.e., normally samples)}

\item{.value}{A column symbol with the value the clustering is based on (e.g., `count`)}

\item{of_samples}{A boolean}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{...}{Further parameters passed to the function kmeans}
}
\value{
A tibble with additional columns
}
\description{
Get SNN shared nearest neighbour clusters to a tibble
}
