% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_differential_transcript_abundance_bulk}
\alias{get_differential_transcript_abundance_bulk}
\title{Get differential transcription information to a tibble using edgeR. At the moment only one covariate is accepted}
\usage{
get_differential_transcript_abundance_bulk(input.df, formula,
  sample_column = NULL, transcript_column = NULL,
  counts_column = NULL, significance_threshold = 0.05)
}
\arguments{
\item{formula}{a formula with no response variable, referring only to numeric variables}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}

\item{df}{A tibble}

\item{design_column}{A character name of the covariate column}
}
\value{
A tibble with edgeR results
}
\description{
Get differential transcription information to a tibble using edgeR. At the moment only one covariate is accepted
}
