% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SE.R
\name{get_differential_transcript_abundance_bulk_voom_SE}
\alias{get_differential_transcript_abundance_bulk_voom_SE}
\title{Get differential transcription information to a tibble using voom.}
\usage{
get_differential_transcript_abundance_bulk_voom_SE(
  .data,
  .formula,
  sample_annotation,
  .contrasts = NULL,
  method = NULL,
  test_above_log2_fold_change = NULL,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = ""
)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{a formula with no response variable, referring only to numeric variables}

\item{.contrasts}{A character vector. See voom makeContrasts specification for the parameter `contrasts`. If contrasts are not present the first covariate is the one the model is tested against (e.g., ~ factor_of_interest)}

\item{method}{A string character. Either "limma_voom", "limma_voom_sample_weights"}

\item{scaling_method}{A character string. The scaling method passed to the backend function (i.e., edgeR::calcNormFactors; "TMM","TMMwsp","RLE","upperquartile")}

\item{omit_contrast_in_colnames}{If just one contrast is specified you can choose to omit the contrast label in the colnames.}
}
\value{
A tibble with voom results
}
\description{
Get differential transcription information to a tibble using voom.
}
\keyword{internal}
