% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SE.R
\name{get_differential_transcript_abundance_deseq2_SE}
\alias{get_differential_transcript_abundance_deseq2_SE}
\title{Get differential transcription information to a tibble using DESeq2}
\usage{
get_differential_transcript_abundance_deseq2_SE(
  .data,
  .formula,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{a formula with no response variable, referring only to numeric variables}

\item{.contrasts}{A character vector. See edgeR makeContrasts specification for the parameter `contrasts`. If contrasts are not present the first covariate is the one the model is tested against (e.g., ~ factor_of_interest)}

\item{method}{A string character. Either "edgeR_quasi_likelihood" (i.e., QLF), "edgeR_likelihood_ratio" (i.e., LRT)}

\item{scaling_method}{A character string. The scaling method passed to the backend function (i.e., edgeR::calcNormFactors; "TMM","TMMwsp","RLE","upperquartile")}

\item{omit_contrast_in_colnames}{If just one contrast is specified you can choose to omit the contrast label in the colnames.}

\item{...}{Additional arguments for DESeq2}
}
\value{
A tibble with edgeR results
}
\description{
Get differential transcription information to a tibble using DESeq2
}
\keyword{internal}
