% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_elements_features}
\alias{get_elements_features}
\title{Get column names either from user or from attributes}
\usage{
get_elements_features(input.df, elements_column, feature_column,
  of_samples)
}
\arguments{
\item{input.df}{A tibble}

\item{elements_column}{A character name of the sample column}

\item{feature_column}{A character name of the gene/transcript name column}
}
\value{
A list of column enquo or error
}
\description{
Get column names either from user or from attributes
}
