% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_ensembl_symbol_mapping}
\alias{get_ensembl_symbol_mapping}
\title{after wget, this function merges hg37 and hg38 mapping data bases - Do not execute!}
\usage{
get_ensembl_symbol_mapping()
}
\value{
A tibble with ensembl-transcript mapping
}
\description{
after wget, this function merges hg37 and hg38 mapping data bases - Do not execute!
}
