% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_rotated_dimensions}
\alias{get_rotated_dimensions}
\title{Get rotated dimensions of two principal components or MDS dimension of choice, of an angle}
\usage{
get_rotated_dimensions(input.df, dimension_1_column, dimension_2_column,
  rotation_degrees, elements_column = NULL, of_samples = T,
  dimension_1_column_rotated = NULL, dimension_2_column_rotated = NULL)
}
\arguments{
\item{input.df}{A tibble}

\item{dimension_1_column}{A character string. The column of the dimension 1}

\item{dimension_2_column}{A character string. The column of the dimension 2}

\item{rotation_degrees}{A real number between 0 and 360}
}
\value{
A tibble with additional rotated columns
}
\description{
Get rotated dimensions of two principal components or MDS dimension of choice, of an angle
}
