% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_sample}
\alias{get_sample}
\title{Get column names either from user or from attributes}
\usage{
get_sample(.data, .sample)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{A character name of the sample column}
}
\value{
A list of column enquo or error
}
\description{
Get column names either from user or from attributes
}
