% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_sample_transcript_counts}
\alias{get_sample_transcript_counts}
\title{Get column names either from user or from attributes}
\usage{
get_sample_transcript_counts(input.df, sample_column, transcript_column,
  counts_column)
}
\arguments{
\item{input.df}{A tibble}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the read count column}
}
\value{
A list of column enquo or error
}
\description{
Get column names either from user or from attributes
}
