% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_scaled_counts_bulk}
\alias{get_scaled_counts_bulk}
\title{Get a tibble with scaled counts using TMM}
\usage{
get_scaled_counts_bulk(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "TMM",
  reference_selection_function = median
)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{method}{A character string. The scaling method passed to the backend function (i.e., edgeR::calcNormFactors; "TMM","TMMwsp","RLE","upperquartile")}

\item{reference_selection_function}{A function between median, mean and max}
}
\value{
A tibble including additional columns
}
\description{
Get a tibble with scaled counts using TMM
}
\keyword{internal}
