% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_transcript}
\alias{get_transcript}
\title{Get column names either from user or from attributes}
\usage{
get_transcript(.data, .transcript)
}
\arguments{
\item{.data}{A tibble}

\item{.transcript}{A character name of the transcript column}
}
\value{
A list of column enquo or error
}
\description{
Get column names either from user or from attributes
}
\keyword{internal}
