% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve_cellularity.R
\name{get_X_cibersort}
\alias{get_X_cibersort}
\title{Get Cibersort reference data}
\usage{
get_X_cibersort()
}
\value{
The X_cibersort reference matrix with genes as rows and cell types as columns
}
\description{
This function loads and returns the X_cibersort reference matrix used for 
cell type deconvolution with the Cibersort and LLSR methods. The reference matrix 
contains gene expression signatures for 22 immune cell types.
}
\examples{
# Get the default Cibersort reference matrix
ref_matrix <- get_X_cibersort()

# Use with deconvolve_cellularity
# se |> deconvolve_cellularity(reference = get_X_cibersort(), method = "cibersort")

}
\references{
Newman, A. M., Liu, C. L., Green, M. R., Gentles, A. J., Feng, W., Xu, Y., Hoang, C. D., Diehn, M., & Alizadeh, A. A. (2015). Robust enumeration of cell subsets from tissue expression profiles. Nature Methods, 12(5), 453-457. doi:10.1038/nmeth.3337
}
