% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_x_y_annotation_columns}
\alias{get_x_y_annotation_columns}
\title{get_x_y_annotation_columns}
\usage{
get_x_y_annotation_columns(
  .data,
  .horizontal,
  .vertical,
  .abundance,
  .abundance_scaled
)
}
\arguments{
\item{.data}{A `tbl` (with at least three columns for sample, feature and transcript abundance) or `SummarizedExperiment` (more convenient if abstracted to tibble with library(tidySummarizedExperiment))}

\item{.horizontal}{The name of the column horizontally presented in the heatmap}

\item{.vertical}{The name of the column vertically presented in the heatmap}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.abundance_scaled}{The name of the transcript/gene scaled abundance column}
}
\value{
A list
}
\description{
This function recognise what are the sample-wise columns and transcrip-wise columns
}
\keyword{internal}
