% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO}
\alias{GO}
\title{GO}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17492 rows and 4 columns.}
\usage{
GO
}
\description{
GO
}
\keyword{datasets}
