% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ifelse2_pipe}
\alias{ifelse2_pipe}
\title{This is a generalisation of ifelse that acceots an object and return an objects}
\usage{
ifelse2_pipe(.x, .p1, .p2, .f1, .f2, .f3 = NULL)
}
\arguments{
\item{input.df}{A tibble}

\item{condition}{A boolean}
}
\value{
A tibble
}
\description{
This is a generalisation of ifelse that acceots an object and return an objects
}
