% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ifelse_pipe}
\alias{ifelse_pipe}
\title{This is a generalisation of ifelse that acceots an object and return an objects}
\usage{
ifelse_pipe(.x, .p, .f1, .f2 = NULL)
}
\arguments{
\item{input.df}{A tibble}

\item{condition}{A boolean}
}
\value{
A tibble
}
\description{
This is a generalisation of ifelse that acceots an object and return an objects
}
