% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{impute_abundance,SummarizedExperiment-method}
\alias{impute_abundance,SummarizedExperiment-method}
\title{impute_abundance}
\usage{
\S4method{impute_abundance}{SummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model where the first covariate is the factor of interest and the second covariate is the unwanted variation (of the kind ~ factor_of_intrest + batch)}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}
}
\value{
A `SummarizedExperiment` object
}
\description{
impute_abundance
}
