% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{inpute_abundance}
\alias{inpute_abundance}
\alias{inpute_abundance,spec_tbl_df-method}
\alias{inpute_abundance,tbl_df-method}
\alias{inpute_abundance,tidybulk-method}
\alias{inpute_abundance,SummarizedExperiment-method}
\alias{inpute_abundance,RangedSummarizedExperiment-method}
\title{inpute transcript abundance if missing from sample-transcript pairs}
\usage{
inpute_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)

\S4method{inpute_abundance}{spec_tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)

\S4method{inpute_abundance}{tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)

\S4method{inpute_abundance}{tidybulk}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)

\S4method{inpute_abundance}{SummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)

\S4method{inpute_abundance}{RangedSummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model where the first covariate is the factor of interest and the second covariate is the unwanted variation (of the kind ~ factor_of_intrest + batch)}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}
}
\value{
A `tbl` non-sparse abundance

A `tbl` with inputed abundnce

A `tbl` with inputed abundnce

A `tbl` with inputed abundnce

A `SummarizedExperiment` object

A `SummarizedExperiment` object
}
\description{
inpute_abundance() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with an edditional adjusted abundance column. This method uses scaled counts if present.
}
\details{
\lifecycle{maturing}

This function inputes the abundance of missing sample-transcript pair using the median of the sample group defined by the formula
}
\examples{


res =
	inpute_abundance(
		tidybulk::counts_mini,
	~ condition,
	.sample = sample,
	.transcript = transcript,
	.abundance = count
)


}
