% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SE.R
\name{keep_variable_transcripts_SE}
\alias{keep_variable_transcripts_SE}
\title{Identify variable genes for dimensionality reduction}
\usage{
keep_variable_transcripts_SE(.data, top = 500, log_transform = TRUE)
}
\arguments{
\item{.data}{A tibble}

\item{top}{An integer. How many top genes to select}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{.sample}{A character name of the sample column}

\item{.transcript}{A character name of the transcript/gene column}

\item{.abundance}{A character name of the read count column}
}
\value{
A tibble filtered genes
}
\description{
Identify variable genes for dimensionality reduction
}
\keyword{internal}
