% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{normalise_counts}
\alias{normalise_counts}
\title{Get a tibble with normalised counts using TMM}
\usage{
normalise_counts(input.df, sample_column = NULL,
  transcript_column = NULL, counts_column = NULL,
  cpm_threshold = 0.5, prop = 3/4, method = "TMM",
  reference_selection_function = median, action = "add", ...)
}
\arguments{
\item{input.df}{A tibble}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}

\item{cpm_threshold}{A real positive number}

\item{prop}{A number between 0 and 1}

\item{method}{A character string}

\item{reference_selection_function}{A fucntion median, or max for the fererence sample}

\item{action}{A character string. Whether to join the new information to the input tibble (add), or just get the non-redundant tibble with the new information (get).}

\item{reference}{A reference matrix, not sure if used anymore}
}
\value{
A tibble including additional columns
}
\description{
Get a tibble with normalised counts using TMM
}
