% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_formula_survival}
\alias{parse_formula_survival}
\title{Formula parser with survival}
\usage{
parse_formula_survival(fm)
}
\arguments{
\item{fm}{A formula}
}
\value{
A character vector
}
\description{
Formula parser with survival
}
