% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{pivot_sample}
\alias{pivot_sample}
\title{Extract sample-wise information}
\usage{
pivot_sample(.data, .sample = NULL)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}
}
\value{
A `tbl` object
}
\description{
pivot_sample() takes as imput a `tbl` formatted as | <SAMPLE> | <ENSEMBL_ID> | <COUNT> | <...> | and returns a `tbl` with only sample-related columns
}
\details{
\lifecycle{maturing}

This functon extracts only sample-related information for downstream analysis (e.g., visualisation). It is disruptive in the sense that it cannot be passed anymore to tidybulk function.
}
\examples{


	pivot_sample(
		tidybulk::counts_mini,
		.sample = sample
	)


}
