% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pivot_sample,tidybulk-method}
\alias{pivot_sample,tidybulk-method}
\title{pivot_sample}
\usage{
\S4method{pivot_sample}{tidybulk}(.data, .sample = NULL)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}
}
\value{
A `tbl` object
}
\description{
pivot_sample
}
