% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{pivot_transcript}
\alias{pivot_transcript}
\title{Extract transcript-wise information}
\usage{
pivot_transcript(.data, .transcript = NULL)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.transcript}{The name of the transcript column}
}
\value{
A `tbl` object
}
\description{
pivot_transcript() takes as imput a `tbl` formatted as | <SAMPLE> | <ENSEMBL_ID> | <COUNT> | <...> | and returns a `tbl` with only sampe-related columns
}
\details{
\lifecycle{maturing}

This functon extracts only transcript-related information for downstream analysis (e.g., visualisation). It is disruptive in the sense that it cannot be passed anymore to tidybulk function.
}
\examples{


	pivot_transcript(
		tidybulk::counts_min,
		.transcript = transcript
	)


}
