% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pivot_transcript,tidybulk-method}
\alias{pivot_transcript,tidybulk-method}
\title{pivot_transcript}
\usage{
\S4method{pivot_transcript}{tidybulk}(.data, .transcript = NULL)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.transcript}{The name of the transcript column}
}
\value{
A `tbl` object
}
\description{
pivot_transcript
}
