% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_reduced_dim}
\alias{plot_reduced_dim}
\title{Plot reduced dimensions}
\usage{
plot_reduced_dim(SE, VAR, method = "PCA", dims = 1:2)
}
\arguments{
\item{SE}{A 'SummarizedExperiment'}

\item{VAR}{A Symbol. What variable in colData to be used for color coding}

\item{method}{A character string. IndicatingMethod of dimensionality reduction to be ploted}

\item{dims}{Vector of dimensions to be plotted}
}
\value{
A ggplot object
}
\description{
Plot reduced dimensions
}
\examples{
## Load airway dataset for examples
data('airway', package = 'airway')
  # Ensure a 'condition' column exists for examples expecting it

    SummarizedExperiment::colData(airway)$condition <- SummarizedExperiment::colData(airway)$dex


counts.PCA =
 airway |>
 identify_abundant() |>
 reduce_dimensions(assay = "counts", method="PCA", .dims = 3)
 
 plot_reduced_dim(counts.PCA, condition, "PCA", 1:2)
}
