% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_normalise_abundance.R
\docType{methods}
\name{quantile_normalise_abundance}
\alias{quantile_normalise_abundance}
\alias{quantile_normalise_abundance,SummarizedExperiment-method}
\alias{quantile_normalise_abundance,RangedSummarizedExperiment-method}
\title{Normalise by quantiles the counts of transcripts/genes}
\usage{
quantile_normalise_abundance(
  .data,
  .abundance = NULL,
  method = "limma_normalize_quantiles",
  target_distribution = NULL
)

\S4method{quantile_normalise_abundance}{SummarizedExperiment}(
  .data,
  .abundance = NULL,
  method = "limma_normalize_quantiles",
  target_distribution = NULL
)

\S4method{quantile_normalise_abundance}{RangedSummarizedExperiment}(
  .data,
  .abundance = NULL,
  method = "limma_normalize_quantiles",
  target_distribution = NULL
)
}
\arguments{
\item{.data}{A `tbl` (with at least three columns for sample, feature and transcript abundance) or `SummarizedExperiment` (more convenient if abstracted to tibble with library(tidySummarizedExperiment))}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{method}{A character string. Either "limma_normalize_quantiles" for limma::normalizeQuantiles or "preprocesscore_normalize_quantiles_use_target" for preprocessCore::normalize.quantiles.use.target for large-scale datasets.}

\item{target_distribution}{A numeric vector. If NULL the target distribution will be calculated by preprocessCore. This argument only affects the "preprocesscore_normalize_quantiles_use_target" method.}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}
}
\value{
A tbl object with additional columns with scaled data as `<NAME OF COUNT COLUMN>_scaled`

A `SummarizedExperiment` object

A `SummarizedExperiment` object
}
\description{
quantile_normalise_abundance() takes as input A `tbl` (with at least three columns for sample, feature and transcript abundance) or `SummarizedExperiment` (more convenient if abstracted to tibble with library(tidySummarizedExperiment)) and Scales transcript abundance compansating for sequencing depth (e.g., with TMM algorithm, Robinson and Oshlack doi.org/10.1186/gb-2010-11-3-r25).
}
\details{
`r lifecycle::badge("maturing")`

Tranform the feature abundance across samples so to have the same quantile distribution (using preprocessCore).

Underlying method

If `limma_normalize_quantiles` is chosen

.data |>limma::normalizeQuantiles()

 If `preprocesscore_normalize_quantiles_use_target` is chosen

.data |>
   preprocessCore::normalize.quantiles.use.target(
      target = preprocessCore::normalize.quantiles.determine.target(.data)
   )
}
\examples{


 tidybulk::se_mini |>
   quantile_normalise_abundance()



}
