% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{reduce_dimensions,spec_tbl_df-method}
\alias{reduce_dimensions,spec_tbl_df-method}
\title{reduce_dimensions}
\usage{
\S4method{reduce_dimensions}{spec_tbl_df}(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  method,
  .dims = 2,
  top = 500,
  of_samples = TRUE,
  log_transform = TRUE,
  scale = TRUE,
  action = "add",
  ...
)
}
\description{
reduce_dimensions
}
