% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{reduce_dimensions}
\alias{reduce_dimensions}
\title{Find reduced dimensions}
\usage{
reduce_dimensions(input.df, value_column, method, components = 1:2,
  elements_column = NULL, feature_column = NULL, of_samples = T,
  log_transform = T, action = "add")
}
\arguments{
\item{input.df}{A tibble}

\item{value_column}{A character string. The column that contains the numeric value (i.e., normally read counts)}

\item{method}{A character string}

\item{components}{A list of integer vectors corresponding to principal components of interest (e.g., list(1:2, 3:4, 5:6))}

\item{elements_column}{A character string. The column that is used to calculate distance (i.e., normally samples)}

\item{feature_column}{A character string. The column that is represents entities to cluster (i.e., normally genes)}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{action}{A character string. Whether to join the new information to the input tibble (add), or just get the non-redundant tibble with the new information (get).}
}
\value{
A tibble with additional columns
}
\description{
Find reduced dimensions
}
