% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R, R/utilities.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{do}
\alias{\%>\%}
\alias{tibble}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{do}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}}
}}

