% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{remove_redundancy,RangedSummarizedExperiment-method}
\alias{remove_redundancy,RangedSummarizedExperiment-method}
\title{remove_redundancy}
\usage{
\S4method{remove_redundancy}{RangedSummarizedExperiment}(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  method,
  of_samples = TRUE,
  correlation_threshold = 0.9,
  top = Inf,
  log_transform = FALSE,
  Dim_a_column = NULL,
  Dim_b_column = NULL
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.element}{The name of the element column (normally samples).}

\item{.feature}{The name of the feature column (normally transcripts/genes)}

\item{.abundance}{The name of the column including the numerical value the clustering is based on (normally transcript abundance)}

\item{method}{A character string. The cluster algorithm to use, ay the moment k-means is the only algorithm included.}

\item{of_samples}{A boolean. In case the input is a tidybulk object, it indicates Whether the element column will be sample or transcript column}

\item{correlation_threshold}{A real number between 0 and 1. For correlation based calculation.}

\item{top}{An integer. How many top genes to select for correlation based method}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{Dim_a_column}{A character string. For reduced_dimension based calculation. The column of one principal component}

\item{Dim_b_column}{A character string. For reduced_dimension based calculation. The column of another principal component}
}
\value{
A `tidybulk` object
}
\description{
remove_redundancy
}
