% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{remove_redundancy,ttBulk-method}
\alias{remove_redundancy,ttBulk-method}
\title{remove_redundancy}
\usage{
\S4method{remove_redundancy}{ttBulk}(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  method,
  of_samples = TRUE,
  correlation_threshold = 0.9,
  top = Inf,
  log_transform = FALSE,
  Dim_a_column,
  Dim_b_column
)
}
\description{
remove_redundancy
}
