% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SE.R
\name{remove_redundancy_elements_though_reduced_dimensions_SE}
\alias{remove_redundancy_elements_though_reduced_dimensions_SE}
\title{Identifies the closest pairs in a MDS context and return one of them}
\usage{
remove_redundancy_elements_though_reduced_dimensions_SE(.data)
}
\arguments{
\item{.data}{A tibble}

\item{of_samples}{A boolean}
}
\value{
A tibble with pairs dropped
}
\description{
Identifies the closest pairs in a MDS context and return one of them
}
\keyword{internal}
