% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SE.R
\name{remove_redundancy_elements_through_correlation_SE}
\alias{remove_redundancy_elements_through_correlation_SE}
\title{Drop redundant elements (e.g., samples) for which feature (e.g., genes) aboundances are correlated}
\usage{
remove_redundancy_elements_through_correlation_SE(
  .data,
  correlation_threshold = 0.9,
  of_samples = TRUE
)
}
\arguments{
\item{.data}{A tibble}

\item{correlation_threshold}{A real number between 0 and 1}

\item{of_samples}{A boolean}

\item{top}{An integer. How many top genes to select}
}
\value{
A tibble with redundant elements removed
}
\description{
Drop redundant elements (e.g., samples) for which feature (e.g., genes) aboundances are correlated
}
\keyword{internal}
