% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_complete_confounders_of_non_interest.R
\docType{methods}
\name{resolve_complete_confounders_of_non_interest,SummarizedExperiment-method}
\alias{resolve_complete_confounders_of_non_interest,SummarizedExperiment-method}
\alias{resolve_complete_confounders_of_non_interest,RangedSummarizedExperiment-method}
\title{resolve_complete_confounders_of_non_interest}
\usage{
\S4method{resolve_complete_confounders_of_non_interest}{SummarizedExperiment}(se, ...)

\S4method{resolve_complete_confounders_of_non_interest}{RangedSummarizedExperiment}(se, ...)
}
\value{
A consistent object (to the input) with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A consistent object (to the input) with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).
}
\description{
resolve_complete_confounders_of_non_interest

resolve_complete_confounders_of_non_interest
}
