% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rotate_dimensions,tbl_df-method}
\alias{rotate_dimensions,tbl_df-method}
\title{rotate_dimensions}
\usage{
\S4method{rotate_dimensions}{tbl_df}(
  .data,
  dimension_1_column,
  dimension_2_column,
  rotation_degrees,
  .element = NULL,
  of_samples = TRUE,
  dimension_1_column_rotated = NULL,
  dimension_2_column_rotated = NULL,
  action = "add"
)
}
\description{
rotate_dimensions
}
