% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rotate_dimensions}
\alias{rotate_dimensions}
\title{Calculate the rotated dimensions, of an arbitrary angle}
\usage{
rotate_dimensions(input.df, dimension_1_column, dimension_2_column,
  rotation_degrees, elements_column = NULL, of_samples = T,
  dimension_1_column_rotated = NULL, dimension_2_column_rotated = NULL,
  action = "add")
}
\arguments{
\item{input.df}{A tibble}

\item{dimension_1_column}{A character string. The column of the dimension 1}

\item{dimension_2_column}{A character string. The column of the dimension 2}

\item{rotation_degrees}{A real number between 0 and 360}

\item{elements_column}{A character string. The column that is used to calculate distance (i.e., normally genes)}

\item{of_samples}{if the input is tt object, this will indicate whether the element column will be sample or transcript column}

\item{dimension_1_column_rotated}{A character string. The column of the rotated dimension 1}

\item{dimension_2_column_rotated}{A character string. The column of the rotated dimension 2}

\item{action}{A character string. Whether to join the new information to the input tibble (add), or just get the non-redundant tibble with the new information (get).}
}
\value{
A tibble with additional rotated columns
}
\description{
Calculate the rotated dimensions, of an arbitrary angle
}
