% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_epic}
\alias{run_epic}
\title{Perform linear equation system analysis through llsr}
\usage{
run_epic(mix, reference = NULL)
}
\arguments{
\item{mix}{A data frame}

\item{reference}{A data frame}
}
\value{
A data frame
}
\description{
Perform linear equation system analysis through llsr
}
\keyword{internal}
