% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{run_llsr}
\alias{run_llsr}
\title{Perform linear equation system analysis through llsr}
\usage{
run_llsr(mix, reference)
}
\arguments{
\item{mix}{A data frame}

\item{reference}{A data frame}
}
\value{
A data frame
}
\description{
Perform linear equation system analysis through llsr
}
\keyword{internal}
