% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{scale_abundance,spec_tbl_df-method}
\alias{scale_abundance,spec_tbl_df-method}
\title{scale_abundance}
\usage{
\S4method{scale_abundance}{spec_tbl_df}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  cpm_threshold = 0.5,
  prop = 3/4,
  method = "TMM",
  reference_selection_function = median,
  action = "add"
)
}
\description{
scale_abundance
}
