% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se}
\alias{se}
\title{SummarizedExperiment}
\format{
An object of class \code{RangedSummarizedExperiment} with 100 rows and 8 columns.
}
\usage{
se
}
\description{
SummarizedExperiment
}
\keyword{datasets}
