% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se_mini}
\alias{se_mini}
\title{SummarizedExperiment mini for vignette}
\format{
An object of class \code{SummarizedExperiment} with 527 rows and 5 columns.
}
\usage{
se_mini
}
\description{
SummarizedExperiment mini for vignette
}
\keyword{datasets}
