% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{symbol_to_entrez}
\alias{symbol_to_entrez}
\title{Get ENTREZ id from gene SYMBOL}
\usage{
symbol_to_entrez(.data, .transcript = NULL, .sample = NULL)
}
\arguments{
\item{.data}{A tt or tbl object.}

\item{.transcript}{A character. The name of the ene symbol column.}

\item{.sample}{The name of the sample column}
}
\description{
Get ENTREZ id from gene SYMBOL
}
