% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_deseq2_df}
\alias{test_deseq2_df}
\title{SummarizedExperiment mini for vignette}
\format{
An object of class \code{DESeqDataSet} with 9921 rows and 7 columns.
}
\usage{
test_deseq2_df
}
\description{
SummarizedExperiment mini for vignette
}
\keyword{datasets}
