% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{test_differential_abundance,spec_tbl_df-method}
\alias{test_differential_abundance,spec_tbl_df-method}
\title{test_differential_abundance}
\usage{
\S4method{test_differential_abundance}{spec_tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .coef = 2,
  .contrasts = NULL,
  significance_threshold = 0.05,
  cpm_threshold = 0.5,
  prop = 3/4,
  fill_missing_values = FALSE,
  action = "add"
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.coef}{An integer. See edgeR specifications}

\item{.contrasts}{A character vector. See edgeR makeContrasts specification for the parameter `contrasts`}

\item{significance_threshold}{A real between 0 and 1 (usually 0.05).}

\item{cpm_threshold}{A real positive number}

\item{prop}{A number between 0 and 1}

\item{fill_missing_values}{A boolean. Whether to fill missing sample/transcript values with the median of the transcript. This is rarely needed.}

\item{action}{A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).}
}
\value{
A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).
}
\description{
test_differential_abundance
}
