% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{test_differential_abundance,tbl_df-method}
\alias{test_differential_abundance,tbl_df-method}
\title{test_differential_abundance}
\usage{
\S4method{test_differential_abundance}{tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .coef = 2,
  .contrasts = NULL,
  significance_threshold = 0.05,
  fill_missing_values = FALSE,
  action = "add"
)
}
\description{
test_differential_abundance
}
