% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{test_differential_abundance}
\alias{test_differential_abundance}
\title{Add differential transcription information to a tbl using edgeR.}
\usage{
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  significance_threshold = 0.05,
  action = "add"
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{significance_threshold}{A real between 0 and 1 (usually 0.05).}

\item{action}{A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).}
}
\value{
A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).
}
\description{
test_differential_abundance() takes as imput a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.
}
\details{
At the moment this function uses edgeR only, but other inference algorithms will be added in the near future.
}
\examples{
\donttest{



	test_differential_abundance(
	    ~ condition,
	    sample,
	    transcript,
	    `count`
	)

}

}
