% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{test_differential_composition}
\alias{test_differential_composition}
\alias{test_differential_composition,spec_tbl_df-method}
\alias{test_differential_composition,tbl_df-method}
\alias{test_differential_composition,tidybulk-method}
\alias{test_differential_composition,SummarizedExperiment-method}
\alias{test_differential_composition,RangedSummarizedExperiment-method}
\title{Add differential tissue composition information to a tbl}
\usage{
test_differential_composition(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  significance_threshold = 0.05
)

\S4method{test_differential_composition}{spec_tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  significance_threshold = 0.05
)

\S4method{test_differential_composition}{tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  significance_threshold = 0.05
)

\S4method{test_differential_composition}{tidybulk}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  significance_threshold = 0.05
)

\S4method{test_differential_composition}{SummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  significance_threshold = 0.05
)

\S4method{test_differential_composition}{RangedSummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  significance_threshold = 0.05
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{method}{A string character. Either "edgeR_quasi_likelihood" (i.e., QLF), "edgeR_likelihood_ratio" (i.e., LRT)}

\item{significance_threshold}{A real between 0 and 1 (usually 0.05).}

\item{action}{A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).}
}
\value{
A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A `SummarizedExperiment` object

A `SummarizedExperiment` object
}
\description{
test_differential_composition() takes as imput a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.
}
\details{
\lifecycle{maturing}

At the moment this function uses edgeR only, but other inference algorithms will be added in the near future.
}
\examples{


	test_differential_composition(
	 tidybulk::counts_mini,
	    ~ condition,
	    sample,
	    transcript,
	    `count`
	)

	# The functon `test_differential_composition` operated with contrasts too

 test_differential_composition(
	    tidybulk::counts_mini,
	    ~ 0 + condition,
	    sample,
	    transcript,
	    `count`,
	    .contrasts = c( "conditionTRUE - conditionFALSE")
 )


}
