% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{test_differential_composition_}
\alias{test_differential_composition_}
\title{Get differential composition information to a tibble using edgeR.}
\usage{
test_differential_composition_(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  significance_threshold = 0.05
)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{a formula with no response variable, referring only to numeric variables}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{method}{A string character. Either "edgeR_quasi_likelihood" (i.e., QLF), "edgeR_likelihood_ratio" (i.e., LRT)}

\item{significance_threshold}{A real between 0 and 1}

\item{minimum_counts}{A positive integer. Minimum counts required for at least some samples.}

\item{minimum_proportion}{A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for scaling procedure.}

\item{fill_missing_values}{A boolean. Whether to fill missing sample/transcript values with the median of the transcript. This is rarely needed.}

\item{scaling_method}{A character string. The scaling method passed to the backend function (i.e., edgeR::calcNormFactors; "TMM","TMMwsp","RLE","upperquartile")}

\item{omit_contrast_in_colnames}{If just one contrast is specified you can choose to omit the contrast label in the colnames.}
}
\value{
A tibble with edgeR results
}
\description{
Get differential composition information to a tibble using edgeR.
}
\keyword{internal}
