% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{test_gene_enrichment,spec_tbl_df-method}
\alias{test_gene_enrichment,spec_tbl_df-method}
\title{test_gene_enrichment}
\usage{
\S4method{test_gene_enrichment}{spec_tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  species,
  cores = 10
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model}

\item{.sample}{The name of the sample column}

\item{.entrez}{The ENTREZ doce of the transcripts/genes}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.contrasts}{= NULL,}

\item{species}{A character. For example, human or mouse}

\item{cores}{An integer. The number of cores available}
}
\value{
A `tbl` object
}
\description{
test_gene_enrichment
}
