% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_gene_enrichment_bulk_EGSEA}
\alias{test_gene_enrichment_bulk_EGSEA}
\title{Get gene enrichment analyses using EGSEA}
\usage{
test_gene_enrichment_bulk_EGSEA(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  method,
  species,
  msigdb.gsets,
  kegg.exclude,
  min.size,
  cores = 10
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model}

\item{.sample}{The name of the sample column}

\item{.entrez}{The ENTREZ code of the transcripts/genes}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.contrasts}{A character vector. See edgeR makeContrasts specification for the parameter `contrasts`. If contrasts are not present the first covariate is the one the model is tested against (e.g., ~ factor_of_interest)}

\item{method}{A character vector. The methods to be included in the ensembl. Type EGSEA::egsea.base() to see the supported GSE methods.}

\item{species}{A character. For example, human or mouse}

\item{cores}{An integer. The number of cores available}
}
\value{
A tibble with edgeR results
}
\description{
Get gene enrichment analyses using EGSEA
}
\keyword{internal}
