% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{test_gene_overrepresentation,spec_tbl_df-method}
\alias{test_gene_overrepresentation,spec_tbl_df-method}
\title{test_gene_overrepresentation}
\usage{
\S4method{test_gene_overrepresentation}{spec_tbl_df}(.data, .sample = NULL, .entrez, .do_test, species)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.entrez}{The ENTREZ ID of the transcripts/genes}

\item{.do_test}{A boolean column name symbol. It indicates the transcript to check}

\item{species}{A character. For example, human or mouse. MSigDB uses the latin species names (e.g., \"Mus musculus\", \"Homo sapiens\")}
}
\value{
A `tbl` object
}
\description{
test_gene_overrepresentation
}
