% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_stratification_cellularity_}
\alias{test_stratification_cellularity_}
\title{Get differential composition information to a tibble using edgeR.}
\usage{
test_stratification_cellularity_(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = NULL,
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{a formula with no response variable, referring only to numeric variables}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{method}{A string character. Either "edgeR_quasi_likelihood" (i.e., QLF), "edgeR_likelihood_ratio" (i.e., LRT)}

\item{reference}{A data frame. The transcript/cell_type data frame of integer transcript abundance}

\item{significance_threshold}{A real between 0 and 1}
}
\value{
A tibble with edgeR results
}
\description{
Get differential composition information to a tibble using edgeR.
}
\keyword{internal}
