% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidybulk,tbl_df-method}
\alias{tidybulk,tbl_df-method}
\title{tidybulk}
\usage{
\S4method{tidybulk}{tbl_df}(.data, .sample, .transcript, .abundance, .abundance_scaled = NULL)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.abundance_scaled}{The name of the transcript/gene scaled abundance column}
}
\value{
A `tidybulk` object
}
\description{
tidybulk
}
