% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidybulk_SAM_BAM,character,character-method}
\alias{tidybulk_SAM_BAM,character,character-method}
\title{tidybulk_SAM_BAM}
\usage{
\S4method{tidybulk_SAM_BAM}{character,character}(file_names, genome = "hg38", ...)
}
\arguments{
\item{file_names}{A character vector}

\item{genome}{A character string}

\item{...}{Further parameters passed to the function Rsubread::featureCounts}
}
\value{
A `tidybulk` object
}
\description{
tidybulk_SAM_BAM
}
