% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{tidybulk_SAM_BAM}
\alias{tidybulk_SAM_BAM}
\alias{tidybulk_SAM_BAM,character,character-method}
\title{Creates a `tt` object from a list of file names of BAM/SAM}
\usage{
tidybulk_SAM_BAM(file_names, genome = "hg38", ...)

\S4method{tidybulk_SAM_BAM}{character,character}(file_names, genome = "hg38", ...)
}
\arguments{
\item{file_names}{A character vector}

\item{genome}{A character string specifying an in-built annotation used for 
read summarization. It has four possible values including "mm10", "mm9", 
"hg38" and "hg19"}

\item{...}{Further parameters passed to the function Rsubread::featureCounts}
}
\value{
A `tidybulk` object

A `tidybulk` object
}
\description{
tidybulk_SAM_BAM() creates a `tt` object from A `tbl` 
  (with at least three columns for sample, feature and transcript abundance) 
  or `SummarizedExperiment` (more convenient if abstracted to tibble with 
  library(tidySummarizedExperiment))
}
\details{
`r lifecycle::badge("maturing")`

This function is based on FeatureCounts package 
  (DOI: 10.1093/bioinformatics/btt656). This function creates a tidybulk 
  object and is useful if you want to avoid to specify .sample, 
  .transcript and .abundance arguments all the times. The tidybulk object 
  have an attribute called internals where these three arguments are stored 
  as metadata. They can be extracted as attr(<object>, "internals").

Underlying core function
Rsubread::featureCounts(annot.inbuilt = genome,nthreads = n_cores, ...)
}
