% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr_methods.R
\name{unnest}
\alias{unnest}
\alias{unnest.default}
\alias{unnest.nested_tidybulk}
\alias{nest}
\alias{nest.default}
\alias{nest.tidybulk}
\title{unnest}
\usage{
unnest(
  .data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique"
)

\method{unnest}{default}(
  .data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique"
)

\method{unnest}{nested_tidybulk}(
  .data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique"
)

nest(.data, ...)

\method{nest}{default}(.data, ...)

\method{nest}{tidybulk}(.data, ...)
}
\arguments{
\item{.data}{A tbl. (See tidyr)}

\item{...}{Name-variable pairs of the form new_col = c(col1, col2, col3) (See tidyr)}
}
\value{
A tt object

A tt object
}
\description{
unnest

nest
}
\examples{

unnest(tidybulk(tidybulk::counts_mini, sample, transcript, count), data = -transcript)


nest(tidybulk(tidybulk::counts_mini, sample, transcript, count), data = -transcript)

}
