% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{ttBulk}
\alias{ttBulk}
\title{Creates a `tt` object from a `tbl``}
\usage{
ttBulk(.data, .sample, .transcript, .abundance)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}
}
\value{
A `ttBulk` object
}
\description{
ttBulk() creates a `tt` object from a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |
}
\details{
This function created a ttBulk object and is useful if you want
to avoid to specify .sample, .transcript and .abundance arguments all the times.
The ttBulk object have an attribute called parameters where these three
arguments are stored as metadata. They can be extracted as attr(<object>, "parameters").
}
\examples{




my_tt =  ttBulk(ttBulk::counts_mini, sample, transcript, count)


}
