% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ttBulk,tbl_df-method}
\alias{ttBulk,tbl_df-method}
\title{ttBulk}
\usage{
\S4method{ttBulk}{tbl_df}(.data, .sample, .transcript, .abundance, .abundance_normalised = NULL)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}
}
\value{
A `ttBulk` object
}
\description{
ttBulk
}
