% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{ttBulk_to_SummarizedExperiment}
\alias{ttBulk_to_SummarizedExperiment}
\title{ttBulk_to_SummarizedExperiment}
\usage{
ttBulk_to_SummarizedExperiment(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}
}
\value{
A SummarizedExperiment
}
\description{
ttBulk_to_SummarizedExperiment
}
