% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tximeta_summarizeToGene_object}
\alias{tximeta_summarizeToGene_object}
\title{Needed for tests tximeta_summarizeToGene_object, It is SummarizedExperiment from tximeta}
\format{
An object of class \code{RangedSummarizedExperiment} with 10 rows and 1 columns.
}
\usage{
tximeta_summarizeToGene_object
}
\description{
Needed for tests tximeta_summarizeToGene_object, It is SummarizedExperiment from tximeta
}
\keyword{datasets}
