% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_manuscript_signature_boxplot}
\alias{vignette_manuscript_signature_boxplot}
\title{Needed for vignette vignette_manuscript_signature_boxplot}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 899 rows and 12 columns.
}
\usage{
vignette_manuscript_signature_boxplot
}
\description{
Needed for vignette vignette_manuscript_signature_boxplot
}
\keyword{datasets}
