% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_manuscript_signature_tsne}
\alias{vignette_manuscript_signature_tsne}
\title{Needed for vignette vignette_manuscript_signature_tsne}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 283 rows and 10 columns.
}
\usage{
vignette_manuscript_signature_tsne
}
\description{
Needed for vignette vignette_manuscript_signature_tsne
}
\keyword{datasets}
