% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_manuscript_signature_tsne2}
\alias{vignette_manuscript_signature_tsne2}
\title{Needed for vignette vignette_manuscript_signature_tsne2}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 283 rows and 9 columns.
}
\usage{
vignette_manuscript_signature_tsne2
}
\description{
Needed for vignette vignette_manuscript_signature_tsne2
}
\keyword{datasets}
