% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_cibersort}
\alias{X_cibersort}
\title{Cibersort reference}
\format{
An object of class \code{data.frame} with 547 rows and 22 columns.
}
\usage{
X_cibersort
}
\description{
Cibersort reference
}
\references{
Newman, A. M., Liu, C. L., Green, M. R., Gentles, A. J., Feng, W., Xu, Y., Hoang, C. D., Diehn, M., & Alizadeh, A. A. (2015). Robust enumeration of cell subsets from tissue expression profiles. Nature Methods, 12(5), 453–457. https://doi.org/10.1038/nmeth.3337
}
\keyword{datasets}
