% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_adjusted_counts_for_unwanted_variation_bulk}
\alias{add_adjusted_counts_for_unwanted_variation_bulk}
\title{Add adjusted count for some batch effect}
\usage{
add_adjusted_counts_for_unwanted_variation_bulk(input.df, formula,
  sample_column = NULL, transcript_column = NULL,
  counts_column = NULL, log_transform = T, ...)
}
\arguments{
\item{input.df}{A tibble}

\item{formula}{a formula with no response variable, of the kind ~ factor_of_intrest + batch}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{...}{Further parameters passed to the function sva::ComBat}
}
\value{
A tibble with adjusted counts
}
\description{
Add adjusted count for some batch effect
}
