% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_cell_type_proportions}
\alias{add_cell_type_proportions}
\title{Add cell type proportions from cibersort}
\usage{
add_cell_type_proportions(input.df, sample_column = NULL,
  transcript_column = NULL, counts_column = NULL, ...)
}
\arguments{
\item{input.df}{A tibble}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}

\item{...}{Further parameters passed to the function Cibersort}
}
\value{
A tibble including additional columns
}
\description{
Add cell type proportions from cibersort
}
