% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_differential_transcript_abundance_bulk}
\alias{add_differential_transcript_abundance_bulk}
\title{Add differential transcription information to a tibble using edgeR.}
\usage{
add_differential_transcript_abundance_bulk(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .coef = 2,
  .contrasts = NULL,
  significance_threshold = 0.05
)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{a formula with no response variable, referring only to numeric variables}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.coef}{An integer. See edgeR specifications}

\item{.contrasts}{A character vector. See edgeR makeContrasts specification for the parameter `contrasts`}

\item{significance_threshold}{A real between 0 and 1}
}
\value{
A tibble with differential_transcript_abundance results
}
\description{
Add differential transcription information to a tibble using edgeR.
}
