% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_normalised_counts_bulk.calcNormFactor}
\alias{add_normalised_counts_bulk.calcNormFactor}
\title{Calculate the norm factor with calcNormFactor from limma}
\usage{
add_normalised_counts_bulk.calcNormFactor(input.df, reference = NULL,
  cpm_threshold = 0.5, prop = 3/4, sample_column = sample,
  transcript_column = transcript, counts_column = count, method)
}
\arguments{
\item{input.df}{A tibble}

\item{reference}{A reference matrix, not sure if used anymore}

\item{cpm_threshold}{A real positive number}

\item{prop}{A number between 0 and 1}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the count column}

\item{method}{A string character}
}
\value{
A list including the filtered data frame and the normalization factors
}
\description{
Calculate the norm factor with calcNormFactor from limma
}
