% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_reduced_dimensions_MDS_bulk}
\alias{add_reduced_dimensions_MDS_bulk}
\title{Add dimensionality information to a tibble using MDS}
\usage{
add_reduced_dimensions_MDS_bulk(input.df, value_column,
  components_list = list(c(1, 2)), elements_column = NULL,
  feature_column = NULL, of_samples = T, log_transform = T)
}
\arguments{
\item{input.df}{A tibble}

\item{components_list}{A list of integer vectors corresponding to principal components of interest (e.g., list(1:2, 3:4, 5:6))}

\item{elements_column}{A character string. The column that is used to calculate distance (i.e., normally samples)}

\item{feature_column}{A character string. The column that is represents entities to cluster (i.e., normally genes)}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{counts_column}{A character string. The column that contains the numeric value (i.e., normally read counts)}
}
\value{
A tibble with additional columns
}
\description{
Add dimensionality information to a tibble using MDS
}
