% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_scaled_counts_bulk.calcNormFactor}
\alias{add_scaled_counts_bulk.calcNormFactor}
\title{Calculate the norm factor with calcNormFactor from limma}
\usage{
add_scaled_counts_bulk.calcNormFactor(
  .data,
  reference = NULL,
  minimum_counts = 0.5,
  prop_threshold = 3/4,
  .sample = sample,
  .transcript = transcript,
  .abundance = count,
  method
)
}
\arguments{
\item{.data}{A tibble}

\item{reference}{A reference matrix, not sure if used anymore}

\item{minimum_counts}{A positive integer. Minimum counts required for at least some samples.}

\item{prop_threshold}{A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for scaling procedure.}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{method}{A string character. The scaling method passed to the backend function (i.e., edgeR::calcNormFactors; "TMM","TMMwsp","RLE","upperquartile")}
}
\value{
A list including the filtered data frame and the normalization factors
}
\description{
Calculate the norm factor with calcNormFactor from limma
}
