% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_scaled_counts_bulk.get_cpm}
\alias{add_scaled_counts_bulk.get_cpm}
\title{Get count per million for TMM scaling.}
\usage{
add_scaled_counts_bulk.get_cpm(
  .data,
  .sample = sample,
  .transcript = transcript,
  .abundance = count,
  minimum_counts = 10
)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{minimum_counts}{A positive integer. Minimum counts required for at least some samples.}
}
\value{
A tibble with an additional column
}
\description{
Get count per million for TMM scaling.
}
