% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{add_scaled_counts_bulk.get_low_expressed}
\alias{add_scaled_counts_bulk.get_low_expressed}
\title{Drop lowly tanscribed genes for TMM normalization}
\usage{
add_scaled_counts_bulk.get_low_expressed(
  .data,
  .sample = sample,
  .transcript = transcript,
  .abundance = count,
  cpm_threshold = 0.5,
  prop_threshold = 3/4
)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{cpm_threshold}{A real positive number. Minimum counts per million required for a selected proportion of samples}

\item{prop_threshold}{A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for scaling procedure.}
}
\value{
A tibble filtered
}
\description{
Drop lowly tanscribed genes for TMM normalization
}
