% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_SE.R
\docType{methods}
\name{adjust_abundance}
\alias{adjust_abundance}
\alias{adjust_abundance,spec_tbl_df-method}
\alias{adjust_abundance,tbl_df-method}
\alias{adjust_abundance,tidybulk-method}
\alias{adjust_abundance,SummarizedExperiment-method}
\alias{adjust_abundance,RangedSummarizedExperiment-method}
\title{Adjust transcript abundance for unwanted variation}
\usage{
adjust_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  log_transform = TRUE,
  action = "add",
  ...
)

\S4method{adjust_abundance}{spec_tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  log_transform = TRUE,
  action = "add",
  ...
)

\S4method{adjust_abundance}{tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  log_transform = TRUE,
  action = "add",
  ...
)

\S4method{adjust_abundance}{tidybulk}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  log_transform = TRUE,
  action = "add",
  ...
)

\S4method{adjust_abundance}{SummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  log_transform = TRUE,
  action = "add",
  ...
)

\S4method{adjust_abundance}{RangedSummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  log_transform = TRUE,
  action = "add",
  ...
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model where the first covariate is the factor of interest and the second covariate is the unwanted variation (of the kind ~ factor_of_intrest + batch)}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{action}{A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).}

\item{...}{Further parameters passed to the function sva::ComBat}
}
\value{
A `tbl` with additional columns for the adjusted counts as `<COUNT COLUMN>_adjusted`

A `tbl` with additional columns for the adjusted counts as `<COUNT COLUMN>_adjusted`

A `tbl` with additional columns for the adjusted counts as `<COUNT COLUMN>_adjusted`

A `tbl` with additional columns for the adjusted counts as `<COUNT COLUMN>_adjusted`

A `SummarizedExperiment` object

A `SummarizedExperiment` object
}
\description{
adjust_abundance() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with an edditional adjusted abundance column. This method uses scaled counts if present.
}
\details{
\lifecycle{maturing}

This function adjusts the abundance for (known) unwanted variation.
At the moment just an unwanted covariated is allowed at a time using Combat (DOI: 10.1093/bioinformatics/bts034)

Underlying method:
	sva::ComBat(data, batch = my_batch,	mod = design,	prior.plots = FALSE, ...)
}
\examples{



cm = tidybulk::counts_mini
cm$batch = 0
cm$batch[cm$sample \%in\% c("SRR1740035", "SRR1740043")] = 1

res =
	adjust_abundance(
		cm,
	~ condition + batch,
	.sample = sample,
	.transcript = transcript,
	.abundance = count
)


}
