% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{aggregate_duplicates}
\alias{aggregate_duplicates}
\title{Aggregates multiple counts from the same samples (e.g., from isoforms), concatenates other character columns, and averages other numeric columns}
\usage{
aggregate_duplicates(.data, aggregation_function = sum, .sample = NULL,
  .transcript = NULL, .abundance = NULL, keep_integer = T)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{aggregation_function}{A function for counts aggregation (e.g., sum,  median, or mean)}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{keep_integer}{A boolean. Whether to force the aggregated counts to integer}
}
\value{
A `tbl` object with aggregated transcript abundance and annotation
}
\description{
aggregate_duplicates() takes as imput a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with aggregated transcripts that were duplicated.
}
\details{
This function aggregates duplicated transcripts (e.g., isoforms, ensembl).
For example, we often have to convert ensembl symbols to gene/transcript symbol,
 but in doing so we have to deal with duplicates. `aggregate_duplicates` takes a tibble
 and column names (as symbols; for `sample`, `transcript` and `count`) as arguments and
 returns a tibble with aggregate transcript with the same name. All the rest of the column
 are appended, and factors and boolean are appended as characters.
}
\examples{
\donttest{



counts \%>\%
    aggregate_duplicates(
    sample,
    transcript,
    `count`,
    aggregation_function = sum
    )

}

}
