% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as_matrix}
\alias{as_matrix}
\title{Get matrix from tibble}
\usage{
as_matrix(tbl, rownames = NULL, do_check = TRUE)
}
\arguments{
\item{tbl}{A tibble}

\item{rownames}{The column name of the input tibble that will become 
the rownames of the output matrix}

\item{do_check}{A boolean}
}
\value{
A matrix
}
\description{
Get matrix from tibble
}
\examples{


tibble(.feature = "CD3G", count=1) |> as_matrix(rownames=.feature)

}
