% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{as_SummarizedExperiment}
\alias{as_SummarizedExperiment}
\alias{as_SummarizedExperiment,tbl_df-method}
\title{as_SummarizedExperiment}
\usage{
as_SummarizedExperiment(.data, .abundance = NULL)

\S4method{as_SummarizedExperiment}{tbl_df}(.data, .abundance = NULL)
}
\arguments{
\item{.data}{A tibble}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}
}
\value{
A `SummarizedExperiment` object

A `SummarizedExperiment` object
}
\description{
as_SummarizedExperiment() creates a `SummarizedExperiment` object from a `tbl` or `tidybulk` tbl formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |
}
