% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{as_SummarizedExperiment}
\alias{as_SummarizedExperiment}
\alias{as_SummarizedExperiment,spec_tbl_df-method}
\alias{as_SummarizedExperiment,tbl_df-method}
\alias{as_SummarizedExperiment,tidybulk-method}
\title{as_SummarizedExperiment}
\usage{
as_SummarizedExperiment(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)

\S4method{as_SummarizedExperiment}{spec_tbl_df}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)

\S4method{as_SummarizedExperiment}{tbl_df}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)

\S4method{as_SummarizedExperiment}{tidybulk}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL
)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}
}
\value{
A `SummarizedExperiment` object

A `SummarizedExperiment` object

A `SummarizedExperiment` object

A `SummarizedExperiment` object
}
\description{
as_SummarizedExperiment() creates a `SummarizedExperiment` 
  object from a `tbl` or `tidybulk` tbl formatted as | <SAMPLE> | 
  <TRANSCRIPT> | <COUNT> | <...> |
}
