% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{cluster_elements}
\alias{cluster_elements}
\title{Get clusters of elements (e.g., samples or transcripts)}
\usage{
cluster_elements(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  method,
  of_samples = TRUE,
  log_transform = TRUE,
  action = "add",
  ...
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.element}{The name of the element column (normally samples).}

\item{.feature}{The name of the feature column (normally transcripts/genes)}

\item{.abundance}{The name of the column including the numerical value the clustering is based on (normally transcript abundance)}

\item{method}{A character string. The cluster algorithm to use, ay the moment k-means is the only algorithm included.}

\item{of_samples}{A boolean. In case the input is a ttBulk object, it indicates Whether the element column will be sample or transcript column}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{action}{A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).}

\item{...}{Further parameters passed to the function kmeans}
}
\value{
A tbl object with additional columns with cluster labels
}
\description{
cluster_elements() takes as imput a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and identify clusters in the data.
}
\details{
\lifecycle{maturing}

identifies clusters in the data, normally of samples.
This function returns a tibble with additional columns for the cluster annotation.
At the moment only k-means clustering is supported, the plan is to introduce more clustering methods.
}
\examples{


    cluster_elements(ttBulk::counts_mini, sample, transcript, count,	centers = 2, method="kmeans")

}
