% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts}
\alias{counts}
\title{Example data set}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1340160 rows and 6 columns.}
\usage{
counts
}
\description{
Example data set
}
\keyword{datasets}
