% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_tt_from_tibble_bulk}
\alias{create_tt_from_tibble_bulk}
\title{Create tt object from tibble}
\usage{
create_tt_from_tibble_bulk(
  .data,
  .sample,
  .transcript,
  .abundance,
  .abundance_scaled = NULL
)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.abundance_scaled}{The name of the transcript/gene scaled abundance column}
}
\value{
A tibble with an additional column
}
\description{
Create tt object from tibble
}
\keyword{internal}
