% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{create_tt_from_tibble_bulk}
\alias{create_tt_from_tibble_bulk}
\title{Create tt object from tibble}
\usage{
create_tt_from_tibble_bulk(input.df, sample_column, transcript_column,
  counts_column)
}
\arguments{
\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the read count column}

\item{df}{A tibble}
}
\value{
A tibble with an additional column
}
\description{
Create tt object from tibble
}
