% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve_cellularity.R
\docType{methods}
\name{deconvolve_cellularity}
\alias{deconvolve_cellularity}
\alias{deconvolve_cellularity,SummarizedExperiment-method}
\alias{deconvolve_cellularity,RangedSummarizedExperiment-method}
\title{Get cell type proportions from samples}
\usage{
deconvolve_cellularity(
  .data,
  .abundance = NULL,
  reference = NULL,
  method = "cibersort",
  prefix = "",
  ...
)

\S4method{deconvolve_cellularity}{SummarizedExperiment}(
  .data,
  .abundance = NULL,
  reference = NULL,
  method = "cibersort",
  prefix = "",
  ...
)

\S4method{deconvolve_cellularity}{RangedSummarizedExperiment}(
  .data,
  .abundance = NULL,
  reference = NULL,
  method = "cibersort",
  prefix = "",
  ...
)
}
\arguments{
\item{.data}{A `tbl` (with at least three columns for sample, feature and transcript abundance) or `SummarizedExperiment` (more convenient if abstracted to tibble with library(tidySummarizedExperiment))}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{reference}{A data frame. The methods cibersort and llsr can accept a custom rectangular dataframe with genes as rows names, cell types as column names and gene-transcript abundance as values. For exampler tidybulk::X_cibersort. The transcript/cell_type data frame of integer transcript abundance. If NULL, the default reference for each algorithm will be used. For llsr will be LM22.}

\item{method}{A character string. The method to be used. Available methods: "cibersort", "llsr", "epic", "mcp_counter", "quantiseq", "xcell". If a vector is provided, an error will be thrown. Default is all available methods.}

\item{prefix}{A character string. The prefix you would like to add to the result columns. It is useful if you want to reshape data.}

\item{...}{Further parameters passed to the function Cibersort}
}
\value{
A consistent object (to the input) including additional columns for each cell type estimated

A `SummarizedExperiment` object

A `SummarizedExperiment` object
}
\description{
deconvolve_cellularity() takes as input A `tbl` (with at least three columns for sample, feature and transcript abundance) or `SummarizedExperiment` (more convenient if abstracted to tibble with library(tidySummarizedExperiment)) and returns a consistent object (to the input) with the estimated cell type abundance for each sample
}
\details{
`r lifecycle::badge("maturing")`

This function infers the cell type composition of our samples
(with the algorithm Cibersort; Newman et al., 10.1038/nmeth.3337).

Underlying method:
CIBERSORT(Y = data, X = reference, ...)
}
\examples{


# Subsetting for time efficiency
tidybulk::se_mini |> deconvolve_cellularity(cores = 1)

}
\references{
Mangiola, S., Molania, R., Dong, R., Doyle, M. A., & Papenfuss, A. T. (2021). tidybulk: an R tidy framework for modular transcriptomic data analysis. Genome Biology, 22(1), 42. doi:10.1186/s13059-020-02233-7

Newman, A. M., Liu, C. L., Green, M. R., Gentles, A. J., Feng, W., Xu, Y., Hoang, C. D., Diehn, M., & Alizadeh, A. A. (2015). Robust enumeration of cell subsets from tissue expression profiles. Nature Methods, 12(5), 453-457. doi:10.1038/nmeth.3337

Racle, J., de Jonge, K., Baumgaertner, P., Speiser, D. E., & Gfeller, D. (2017). Simultaneous enumeration of cancer and immune cell types from bulk tumor gene expression data. eLife, 6, e26476. doi:10.7554/eLife.26476
}
