% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{deconvolve_cellularity,spec_tbl_df-method}
\alias{deconvolve_cellularity,spec_tbl_df-method}
\title{deconvolve_cellularity}
\usage{
\S4method{deconvolve_cellularity}{spec_tbl_df}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  reference = X_cibersort,
  method = "cibersort",
  action = "add",
  ...
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{reference}{A data frame. The transcript/cell_type data frame of integer transcript abundance}

\item{method}{A character string. The method to be used. At the moment Cibersort (default) and llsr (linear least squares regression) are available.}

\item{action}{A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).}

\item{...}{Further parameters passed to the function Cibersort}
}
\value{
A `tbl` object including additional columns for each cell type estimated
}
\description{
deconvolve_cellularity
}
