% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_transcript.R
\docType{methods}
\name{describe_transcript}
\alias{describe_transcript}
\alias{.describe_transcript_SE}
\alias{describe_transcript,SummarizedExperiment-method}
\alias{describe_transcript,RangedSummarizedExperiment-method}
\title{Get DESCRIPTION from gene SYMBOL for Human and Mouse}
\usage{
describe_transcript(.data)

.describe_transcript_SE(.data)

\S4method{describe_transcript}{SummarizedExperiment}(.data)

\S4method{describe_transcript}{RangedSummarizedExperiment}(.data)
}
\arguments{
\item{.data}{A tt or tbl object.}
}
\value{
A tbl

A `SummarizedExperiment` object

A consistent object (to the input) including additional columns for transcript symbol

A consistent object (to the input) including additional columns for transcript symbol
}
\description{
Get DESCRIPTION from gene SYMBOL for Human and Mouse

describe_transcript

describe_transcript

describe_transcript
}
\examples{
## Load airway dataset for examples

  data('airway', package = 'airway')
  # Ensure a 'condition' column exists for examples expecting it

    SummarizedExperiment::colData(airway)$condition <- SummarizedExperiment::colData(airway)$dex



describe_transcript(airway)

}
\references{
Mangiola, S., Molania, R., Dong, R., Doyle, M. A., & Papenfuss, A. T. (2021). tidybulk: an R tidy framework for modular transcriptomic data analysis. Genome Biology, 22(1), 42. doi:10.1186/s13059-020-02233-7

Carlson, M. (2019). org.Hs.eg.db: Genome wide annotation for Human. R package version 3.8.2.

Carlson, M. (2019). org.Mm.eg.db: Genome wide annotation for Mouse. R package version 3.8.2.
}
