% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_transcript.R
\docType{methods}
\name{describe_transcript}
\alias{describe_transcript}
\alias{.describe_transcript_SE}
\alias{describe_transcript,SummarizedExperiment-method}
\alias{describe_transcript,RangedSummarizedExperiment-method}
\title{Get DESCRIPTION from gene SYMBOL for Human and Mouse}
\usage{
describe_transcript(.data)

.describe_transcript_SE(.data)

\S4method{describe_transcript}{SummarizedExperiment}(.data)

\S4method{describe_transcript}{RangedSummarizedExperiment}(.data)
}
\arguments{
\item{.data}{A tt or tbl object.}

\item{.transcript}{A character. The name of the gene symbol column.}
}
\value{
A tbl

A `SummarizedExperiment` object

A consistent object (to the input) including additional columns for transcript symbol

A consistent object (to the input) including additional columns for transcript symbol
}
\description{
Get DESCRIPTION from gene SYMBOL for Human and Mouse

describe_transcript

describe_transcript

describe_transcript
}
\examples{

describe_transcript(tidybulk::se_mini)

}
