% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{drop_redundant_elements_through_correlation}
\alias{drop_redundant_elements_through_correlation}
\title{Drop redundant elements (e.g., samples) for which feature (e.g., genes) aboundances are correlated}
\usage{
drop_redundant_elements_through_correlation(input.df, value_column,
  correlation_threshold = 0.9, elements_column = NULL,
  feature_column = NULL, of_samples = T, log_transform = F)
}
\arguments{
\item{input.df}{A tibble}

\item{correlation_threshold}{A real number between 0 and 1}

\item{elements_column}{A character string. The column that is used to calculate distance (i.e., normally samples)}

\item{feature_column}{A character string. The column that is represents entities to cluster (i.e., normally genes)}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{counts_column}{A character string. The column that contains the numeric value (i.e., normally counts)}
}
\value{
A tibble with redundant elemens removed
}
\description{
Drop redundant elements (e.g., samples) for which feature (e.g., genes) aboundances are correlated
}
