% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{filter_abundant}
\alias{filter_abundant}
\title{Filter abundant transcripts}
\usage{
filter_abundant(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  cpm_threshold = 0.5,
  prop_threshold = 3/4
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{cpm_threshold}{A real positive number. It is the threshold of count per million that is used to filter transcripts/genes out from the normalisation procedure.}

\item{prop_threshold}{A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for normalisation procedure.}
}
\value{
A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).
}
\description{
filter_abundant() takes as imput a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.
}
\details{
\lifecycle{maturing}

At the moment this function uses edgeR only, but other inference algorithms will be added in the near future.
}
\examples{



	filter_abundant(
	ttBulk::counts_mini,
	    sample,
	    transcript,
	    `count`
	)


}
