% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{filter_abundant,SummarizedExperiment-method}
\alias{filter_abundant,SummarizedExperiment-method}
\title{filter_abundant}
\usage{
\S4method{filter_abundant}{SummarizedExperiment}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  cpm_threshold = 0.5,
  prop_threshold = 3/4
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{cpm_threshold}{A real positive number. It is the threshold of count per million that is used to filter transcripts/genes out from the scaling procedure.}

\item{prop_threshold}{A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for scaling procedure.}
}
\value{
A `tidyBulk` object
}
\description{
filter_abundant
}
