% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_adjusted_counts_for_unwanted_variation_bulk}
\alias{get_adjusted_counts_for_unwanted_variation_bulk}
\title{Get adjusted count for some batch effect}
\usage{
get_adjusted_counts_for_unwanted_variation_bulk(.data, .formula,
  .sample = NULL, .transcript = NULL, .abundance = NULL,
  log_transform = T, ...)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{a formula with no response variable, of the kind ~ factor_of_intrest + batch}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{...}{Further parameters passed to the function sva::ComBat}
}
\value{
A tibble with adjusted counts
}
\description{
Get adjusted count for some batch effect
}
