% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_adjusted_counts_for_unwanted_variation_bulk}
\alias{get_adjusted_counts_for_unwanted_variation_bulk}
\title{Get adjusted read count for some batch effect}
\usage{
get_adjusted_counts_for_unwanted_variation_bulk(input.df, formula,
  sample_column = NULL, transcript_column = NULL,
  counts_column = NULL, log_transform = T)
}
\arguments{
\item{formula}{a formula with no response variable, of the kind ~ factor_of_intrest + batch}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the read count column}

\item{df}{A tibble}

\item{design_column}{A character name of the covariate column}
}
\value{
A tibble with adjusted counts
}
\description{
Get adjusted read count for some batch effect
}
