% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_bibliography}
\alias{get_bibliography}
\alias{get_bibliography,tidybulk-method}
\title{Produces the bibliography list of your workflow}
\usage{
get_bibliography(.data)

\S4method{get_bibliography}{tidybulk}(.data)
}
\arguments{
\item{.data}{A `tidybulk` tibble}
}
\value{
A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).
}
\description{
get_bibliography() takes as input a `tidybulk`
}
\details{
\lifecycle{maturing}

This methods returns the bibliography list of your workflow from the internals of a tidybulk tibble (attr(., "internals"))
}
\examples{

# Define tidybulk tibble
df = tidybulk(tidybulk::counts_mini, sample, transcript, count)

get_bibliography(df)



}
