% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_SE.R
\docType{methods}
\name{get_bibliography}
\alias{get_bibliography}
\alias{get_bibliography,tbl-method}
\alias{get_bibliography,tbl_df-method}
\alias{get_bibliography,spec_tbl_df-method}
\alias{get_bibliography,tidybulk-method}
\alias{get_bibliography,SummarizedExperiment-method}
\alias{get_bibliography,RangedSummarizedExperiment-method}
\title{Produces the bibliography list of your workflow}
\usage{
get_bibliography(.data)

\S4method{get_bibliography}{tbl}(.data)

\S4method{get_bibliography}{tbl_df}(.data)

\S4method{get_bibliography}{spec_tbl_df}(.data)

\S4method{get_bibliography}{tidybulk}(.data)

\S4method{get_bibliography}{SummarizedExperiment}(.data)

\S4method{get_bibliography}{RangedSummarizedExperiment}(.data)
}
\arguments{
\item{.data}{A `tbl` (with at least three columns for sample, feature 
and transcript abundance) or `SummarizedExperiment` (more convenient if 
abstracted to tibble with library(tidySummarizedExperiment))}
}
\value{
NULL. It prints a list of bibliography references for the software 
used through the workflow.

A consistent object (to the input) with additional columns for the 
statistics from the hypothesis test (e.g.,  log fold change, p-value and 
false discovery rate).

A consistent object (to the input) with additional columns for the 
statistics from the hypothesis test (e.g.,  log fold change, p-value and 
false discovery rate).
}
\description{
get_bibliography() takes as input a `tidybulk`
}
\details{
`r lifecycle::badge("maturing")`

This methods returns the bibliography list of your workflow from 
the internals of a tidybulk object (attr(., "internals"))
}
\examples{
get_bibliography(tidybulk::se_mini)


}
