% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_cell_type_proportions}
\alias{get_cell_type_proportions}
\title{Get cell type proportions from cibersort}
\usage{
get_cell_type_proportions(input.df, sample_column = NULL,
  transcript_column = NULL, counts_column = NULL, ...)
}
\arguments{
\item{input.df}{A tibble}

\item{sample_column}{A character name of the sample column}

\item{transcript_column}{A character name of the gene/transcript name column}

\item{counts_column}{A character name of the read count column}
}
\value{
A tibble including additional columns
}
\description{
Get cell type proportions from cibersort
}
