% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_cell_type_proportions}
\alias{get_cell_type_proportions}
\title{Get cell type proportions from cibersort}
\usage{
get_cell_type_proportions(.data, .sample = NULL, .transcript = NULL,
  .abundance = NULL, ...)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{...}{Further parameters passed to the function Cibersort}
}
\value{
A tibble including additional columns
}
\description{
Get cell type proportions from cibersort
}
