% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_differential_transcript_abundance_bulk}
\alias{get_differential_transcript_abundance_bulk}
\title{Get differential transcription information to a tibble using edgeR.}
\usage{
get_differential_transcript_abundance_bulk(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  significance_threshold = 0.05
)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{a formula with no response variable, referring only to numeric variables}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{significance_threshold}{A real between 0 and 1}
}
\value{
A tibble with edgeR results
}
\description{
Get differential transcription information to a tibble using edgeR.
}
