% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_differential_transcript_abundance_bulk}
\alias{get_differential_transcript_abundance_bulk}
\title{Get differential transcription information to a tibble using edgeR.}
\usage{
get_differential_transcript_abundance_bulk(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  significance_threshold = 0.05,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  fill_missing_values = FALSE,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE
)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{a formula with no response variable, referring only to numeric variables}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.contrasts}{A character vector. See edgeR makeContrasts specification for the parameter `contrasts`. If contrasts are not present the first covariate is the one the model is tested against (e.g., ~ factor_of_interest)}

\item{significance_threshold}{A real between 0 and 1}

\item{minimum_counts}{A positive integer. Minimum counts required for at least some samples.}

\item{minimum_proportion}{A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for scaling procedure.}

\item{fill_missing_values}{A boolean. Whether to fill missing sample/transcript values with the median of the transcript. This is rarely needed.}

\item{scaling_method}{A character string. The scaling method passed to the backend function (i.e., edgeR::calcNormFactors; "TMM","TMMwsp","RLE","upperquartile")}

\item{omit_contrast_in_colnames}{If just one contrast is specified you can choose to omit the contrast label in the colnames.}
}
\value{
A tibble with edgeR results
}
\description{
Get differential transcription information to a tibble using edgeR.
}
