% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_elements}
\alias{get_elements}
\title{Get column names either from user or from attributes}
\usage{
get_elements(.data, .element)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A character name of the sample column}
}
\value{
A list of column enquo or error
}
\description{
Get column names either from user or from attributes
}
