% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_normalised_counts_bulk}
\alias{get_normalised_counts_bulk}
\title{Get a tibble with normalised counts using TMM}
\usage{
get_normalised_counts_bulk(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  cpm_threshold = 0.5,
  prop = 3/4,
  method = "TMM",
  reference_selection_function = median
)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{cpm_threshold}{A real positive number}

\item{prop}{A number between 0 and 1}

\item{method}{A character string}

\item{reference_selection_function}{A function between median, mean and max}
}
\value{
A tibble including additional columns
}
\description{
Get a tibble with normalised counts using TMM
}
