% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_reduced_dimensions_PCA_bulk}
\alias{get_reduced_dimensions_PCA_bulk}
\title{Get principal component information to a tibble using PCA}
\usage{
get_reduced_dimensions_PCA_bulk(.data, .value, components = 1:2,
  .element = NULL, .feature = NULL, of_samples = T,
  log_transform = T, scale = T, ...)
}
\arguments{
\item{.data}{A tibble}

\item{.value}{A column symbol with the value the clustering is based on (e.g., `count`)}

\item{components}{A integer vector corresponding to principal components of interest (e.g., 1:6)}

\item{.element}{A column symbol. The column that is used to calculate distance (i.e., normally samples)}

\item{.feature}{A column symbol. The column that is represents entities to cluster (i.e., normally genes)}

\item{of_samples}{A boolean}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}

\item{scale}{A boolean}

\item{...}{Further parameters passed to the function prcomp}
}
\value{
A tibble with additional columns
}
\description{
Get principal component information to a tibble using PCA
}
