% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_symbol_from_ensembl}
\alias{get_symbol_from_ensembl}
\title{Get transcript column from ensembl gene id}
\usage{
get_symbol_from_ensembl(input.df, ensembl_transcript_column)
}
\arguments{
\item{input.df}{A tibble}

\item{ensembl_transcript_column}{A character string. The column that is represents ensembl gene id}
}
\value{
A tibble with added annotation
}
\description{
Get transcript column from ensembl gene id
}
