% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_SE.R
\docType{methods}
\name{identify_abundant}
\alias{identify_abundant}
\alias{identify_abundant,spec_tbl_df-method}
\alias{identify_abundant,tbl_df-method}
\alias{identify_abundant,tidybulk-method}
\alias{identify_abundant,SummarizedExperiment-method}
\alias{identify_abundant,RangedSummarizedExperiment-method}
\title{Identify abundant transcripts/genes}
\usage{
identify_abundant(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  design = NULL,
  formula_design = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  minimum_count_per_million = NULL,
  ...
)

\S4method{identify_abundant}{spec_tbl_df}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  design = NULL,
  formula_design = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  minimum_count_per_million = NULL
)

\S4method{identify_abundant}{tbl_df}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  design = NULL,
  formula_design = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  minimum_count_per_million = NULL
)

\S4method{identify_abundant}{tidybulk}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  design = NULL,
  formula_design = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  minimum_count_per_million = NULL
)

\S4method{identify_abundant}{SummarizedExperiment}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  design = NULL,
  formula_design = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  minimum_count_per_million = NULL
)

\S4method{identify_abundant}{RangedSummarizedExperiment}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  design = NULL,
  formula_design = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  minimum_count_per_million = NULL
)
}
\arguments{
\item{.data}{A `tbl` or `SummarizedExperiment` object containing transcript/gene abundance data}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{factor_of_interest}{The name of the column containing groups/conditions for filtering. 
Used by edgeR's filterByExpr to define sample groups.}

\item{design}{A design matrix for more complex experimental designs. If provided, 
this is passed to filterByExpr instead of factor_of_interest.}

\item{formula_design}{A formula to construct the design matrix from colData. If provided, overrides the design argument. Default is NULL.}

\item{minimum_counts}{A positive number specifying the minimum counts per million (CPM) threshold
for a transcript to be considered abundant (default = 10)}

\item{minimum_proportion}{A number between 0 and 1 specifying the minimum proportion of samples
that must exceed the minimum_counts threshold (default = 0.7)}

\item{minimum_count_per_million}{Minimum CPM cutoff to use for filtering (passed to CPM.Cutoff in filterByExpr). If provided, this will override the minimum_counts parameter. Default is NULL (uses edgeR default).}
}
\value{
Returns the input object with an additional logical column `.abundant` indicating which 
features passed the abundance threshold criteria.

A consistent object (to the input) with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A consistent object (to the input) with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A consistent object (to the input) with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A `SummarizedExperiment` object

A `SummarizedExperiment` object
}
\description{
Identifies transcripts/genes that are consistently expressed above a threshold across samples. 
This function adds a logical column `.abundant` to indicate which features pass the filtering criteria.
}
\details{
`r lifecycle::badge("maturing")`


This function uses edgeR's filterByExpr() function to identify consistently expressed features.
A feature is considered abundant if it has CPM > minimum_counts in at least minimum_proportion 
of samples in at least one experimental group (defined by factor_of_interest or design).
If minimum_count_per_million is provided, it will override the minimum_counts parameter.
}
\examples{
# Basic usage
se_mini |> identify_abundant()

# With custom thresholds
se_mini |> identify_abundant(
  minimum_counts = 5,
  minimum_proportion = 0.5
)

# Using a factor of interest
se_mini |> identify_abundant(factor_of_interest = condition)

}
\references{
McCarthy, D. J., Chen, Y., & Smyth, G. K. (2012). Differential expression analysis of 
multifactor RNA-Seq experiments with respect to biological variation. Nucleic Acids Research, 
40(10), 4288-4297. DOI: 10.1093/bioinformatics/btp616
}
