% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{keep_abundant,SummarizedExperiment-method}
\alias{keep_abundant,SummarizedExperiment-method}
\title{keep_abundant}
\usage{
\S4method{keep_abundant}{SummarizedExperiment}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{factor_of_interest}{The name of the column of the factor of interest. This is used for defining sample groups for the filtering process.}

\item{minimum_counts}{A real positive number. It is the threshold of count per million that is used to filter transcripts/genes out from the scaling procedure.}

\item{minimum_proportion}{A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for scaling procedure.}
}
\value{
A `tidybulk` object
}
\description{
keep_abundant
}
