% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{merged_ttBulk_object}
\alias{merged_ttBulk_object}
\title{Add cell type information on single cells}
\usage{
merged_ttBulk_object(...)
}
\arguments{
\item{...}{Argument to pass to bind_rows}
}
\value{
A tt tt object
}
\description{
Add cell type information on single cells
}
