% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr_methods.R
\name{nest}
\alias{nest}
\title{nest}
\usage{
nest(.data, ...)
}
\arguments{
\item{.data}{A tbl. (See tidyr)}

\item{...}{Name-variable pairs of the form new_col = c(col1, col2, col3) (See tidyr)}
}
\value{
A tt object
}
\description{
nest
}
\examples{

nest(tidybulk(tidybulk::counts_mini, sample, transcript, count), data = -transcript)


}
