% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_SE.R
\docType{methods}
\name{pivot_sample}
\alias{pivot_sample}
\alias{pivot_sample,spec_tbl_df-method}
\alias{pivot_sample,tbl_df-method}
\alias{pivot_sample,tidybulk-method}
\alias{pivot_sample,SummarizedExperiment-method}
\alias{pivot_sample,RangedSummarizedExperiment-method}
\title{Extract sample-wise information}
\usage{
pivot_sample(.data, .sample = NULL)

\S4method{pivot_sample}{spec_tbl_df}(.data, .sample = NULL)

\S4method{pivot_sample}{tbl_df}(.data, .sample = NULL)

\S4method{pivot_sample}{tidybulk}(.data, .sample = NULL)

\S4method{pivot_sample}{SummarizedExperiment}(.data, .sample = NULL)

\S4method{pivot_sample}{RangedSummarizedExperiment}(.data, .sample = NULL)
}
\arguments{
\item{.data}{A `tbl` (with at least three columns for sample, feature and 
transcript abundance) or `SummarizedExperiment` (more convenient if 
abstracted to tibble with library(tidySummarizedExperiment))}

\item{.sample}{The name of the sample column}
}
\value{
A `tbl` with transcript-related information

A consistent object (to the input)

A consistent object (to the input)
}
\description{
pivot_sample() takes as input a `tbl` 
(with at least three columns for sample, feature and transcript abundance) 
or `SummarizedExperiment` (more convenient if abstracted to tibble with 
library(tidySummarizedExperiment)) and returns a `tbl` with only sample-related columns
}
\details{
`r lifecycle::badge("maturing")`

This functon extracts only sample-related information for 
downstream analysis (e.g., visualisation). It is disruptive in the sense 
that it cannot be passed anymore to tidybulk function.
}
\examples{


	pivot_sample(tidybulk::se_mini )


}
