% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_SE.R
\docType{methods}
\name{pivot_transcript}
\alias{pivot_transcript}
\alias{pivot_transcript,spec_tbl_df-method}
\alias{pivot_transcript,tbl_df-method}
\alias{pivot_transcript,tidybulk-method}
\alias{pivot_transcript,SummarizedExperiment-method}
\alias{pivot_transcript,RangedSummarizedExperiment-method}
\title{Extract transcript-wise information}
\usage{
pivot_transcript(.data, .transcript = NULL)

\S4method{pivot_transcript}{spec_tbl_df}(.data, .transcript = NULL)

\S4method{pivot_transcript}{tbl_df}(.data, .transcript = NULL)

\S4method{pivot_transcript}{tidybulk}(.data, .transcript = NULL)

\S4method{pivot_transcript}{SummarizedExperiment}(.data, .transcript = NULL)

\S4method{pivot_transcript}{RangedSummarizedExperiment}(.data, .transcript = NULL)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.transcript}{The name of the transcript column}
}
\value{
A `tbl` object

A `tbl` object

A `tbl` object

A `tbl` object

A `tbl` object

A `tbl` object
}
\description{
pivot_transcript() takes as input a `tbl` formatted as | <SAMPLE> | <ENSEMBL_ID> | <COUNT> | <...> | and returns a `tbl` with only sample-related columns
}
\details{
`r lifecycle::badge("maturing")`

This functon extracts only transcript-related information for downstream analysis (e.g., visualisation). It is disruptive in the sense that it cannot be passed anymore to tidybulk function.
}
\examples{


	pivot_transcript(
		tidybulk::counts_mini,
		.transcript = transcript
	)


}
