% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{remove_redundancy_elements_through_correlation}
\alias{remove_redundancy_elements_through_correlation}
\title{Drop redundant elements (e.g., samples) for which feature (e.g., genes) aboundances are correlated}
\usage{
remove_redundancy_elements_through_correlation(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  correlation_threshold = 0.9,
  top = Inf,
  of_samples = TRUE,
  log_transform = FALSE
)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A column symbol. The column that is used to calculate distance (i.e., normally samples)}

\item{.feature}{A column symbol. The column that is represents entities to cluster (i.e., normally genes)}

\item{.abundance}{A column symbol with the value the clustering is based on (e.g., `count`)}

\item{correlation_threshold}{A real number between 0 and 1}

\item{top}{An integer. How many top genes to select}

\item{of_samples}{A boolean}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}
}
\value{
A tibble with redundant elements removed
}
\description{
Drop redundant elements (e.g., samples) for which feature (e.g., genes) aboundances are correlated
}
\keyword{internal}
