% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_SE.R
\docType{methods}
\name{resolve_complete_confounders_of_non_interest}
\alias{resolve_complete_confounders_of_non_interest}
\alias{resolve_complete_confounders_of_non_interest,SummarizedExperiment-method}
\alias{resolve_complete_confounders_of_non_interest,RangedSummarizedExperiment-method}
\title{Resolve Complete Confounders of Non-Interest}
\usage{
resolve_complete_confounders_of_non_interest(se, ...)

\S4method{resolve_complete_confounders_of_non_interest}{SummarizedExperiment}(se, ...)

\S4method{resolve_complete_confounders_of_non_interest}{RangedSummarizedExperiment}(se, ...)
}
\arguments{
\item{se}{A SummarizedExperiment object that contains the data to be processed.}

\item{...}{Arbitrary number of factor variables represented as symbols or quosures
to be considered for resolving confounders. These factors are processed
in combinations of two.}
}
\value{
A modified SummarizedExperiment object with confounders resolved.

A consistent object (to the input) with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A consistent object (to the input) with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).
}
\description{
This generic function processes a SummarizedExperiment object to handle confounders
that are not of interest in the analysis. It dynamically handles combinations
of provided factors, adjusting the data by nesting and summarizing over these factors.
}
\examples{
# Not run:
# se is a SummarizedExperiment object
# resolve_complete_confounders_of_non_interest(se, factor1, factor2, factor3)
}
