% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{scale_abundance,ttBulk-method}
\alias{scale_abundance,ttBulk-method}
\title{scale_abundance}
\usage{
\S4method{scale_abundance}{ttBulk}(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  cpm_threshold = 0.5,
  prop = 3/4,
  method = "TMM",
  reference_selection_function = median,
  action = "add"
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{cpm_threshold}{A real positive number. It is the threshold of count per million that is used to filter transcripts/genes out from the normalisation procedure. The normalisation inference is then applied back to all unfiltered data.}

\item{prop}{A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for normalisation procedure.}

\item{method}{A character string. The methods used by the function. The method is passed to the function `calcNormFactors` from edgeR package.}

\item{reference_selection_function}{A fucntion that is used to selecting the reference sample for normalisation. It could be max (default), which choose the sample with maximum library size; or median, which chooses the sample with median library size.}

\item{action}{A character string between "add" (default) and "get". "add" joins the new information to the input tbl (default), "get" return a non-redundant tbl with the just new information.}
}
\value{
A tbl object with additional columns with scaled data as `<NAME OF COUNT COLUMN> scaled`
}
\description{
scale_abundance
}
