% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{symbol_to_entrez}
\alias{symbol_to_entrez}
\title{Get ENTREZ id from gene SYMBOL}
\usage{
symbol_to_entrez(.data, .transcript = NULL, .sample = NULL)
}
\arguments{
\item{.data}{A tt or tbl object.}

\item{.transcript}{A character. The name of the gene symbol column.}

\item{.sample}{The name of the sample column}
}
\value{
A tbl
}
\description{
Get ENTREZ id from gene SYMBOL
}
\examples{

# This function was designed for data.frame
# Convert from SummarizedExperiment for this example. It is NOT reccomended.

tidybulk::se_mini |> tidybulk() |> as_tibble() |> 
symbol_to_entrez(.transcript = .feature, .sample = .sample)

}
