% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{test_differential_abundance}
\alias{test_differential_abundance}
\title{Add differential transcription information to a tbl using edgeR.}
\usage{
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .coef = 2,
  .contrasts = NULL,
  significance_threshold = 0.05,
  cpm_threshold = 0.5,
  prop_threshold = 3/4,
  fill_missing_values = FALSE,
  action = "add"
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.coef}{An integer. See edgeR specifications}

\item{.contrasts}{A character vector. See edgeR makeContrasts specification for the parameter `contrasts`}

\item{significance_threshold}{A real between 0 and 1 (usually 0.05).}

\item{cpm_threshold}{A real positive number. It is the threshold of count per million that is used to filter transcripts/genes out from the normalisation procedure.}

\item{prop_threshold}{A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for normalisation procedure.}

\item{fill_missing_values}{A boolean. Whether to fill missing sample/transcript values with the median of the transcript. This is rarely needed.}

\item{action}{A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).}
}
\value{
A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).
}
\description{
test_differential_abundance() takes as imput a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.
}
\details{
\lifecycle{maturing}

At the moment this function uses edgeR only, but other inference algorithms will be added in the near future.
}
\examples{


	test_differential_abundance(
	 tidyBulk::counts_mini,
	    ~ condition,
	    sample,
	    transcript,
	    `count`
	)

	# The functon `test_differential_abundance` operated with contrasts too

 test_differential_abundance(
	    tidyBulk::counts_mini,
	    ~ 0 + condition,
	    sample,
	    transcript,
	    `count`,
	    .contrasts = c( "conditionTRUE - conditionFALSE")
 )


}
