% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{test_differential_cellularity}
\alias{test_differential_cellularity}
\alias{test_differential_cellularity,spec_tbl_df-method}
\alias{test_differential_cellularity,tbl_df-method}
\alias{test_differential_cellularity,tidybulk-method}
\alias{test_differential_cellularity,SummarizedExperiment-method}
\alias{test_differential_cellularity,RangedSummarizedExperiment-method}
\title{Add differential tissue composition information to a tbl}
\usage{
test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

\S4method{test_differential_cellularity}{spec_tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

\S4method{test_differential_cellularity}{tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

\S4method{test_differential_cellularity}{tidybulk}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

\S4method{test_differential_cellularity}{SummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

\S4method{test_differential_cellularity}{RangedSummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model. If censored regression is desired (coxph) the formula should be of the form \"survival::Surv\(y, dead\) ~ x\"}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{method}{A string character. Either \"cibersort\" or \"llsr\"}

\item{reference}{A data frame. The transcript/cell_type data frame of integer transcript abundance}

\item{significance_threshold}{A real between 0 and 1 (usually 0.05).}

\item{...}{Further parameters passed to the method deconvolve_cellularity}
}
\value{
A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g.,  log fold change, p-value and false discovery rate).

A `SummarizedExperiment` object

A `SummarizedExperiment` object
}
\description{
test_differential_cellularity() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.
}
\details{
\lifecycle{maturing}

This routine applies a deconvolution method (e.g., Cibersort; DOI: 10.1038/nmeth.3337)
and passes the proportions inferred into a generalised linear model (DOI:dx.doi.org/10.1007/s11749-010-0189-z)
or a cox regression model (ISBN: 978-1-4757-3294-8)

Underlying method for the generalised linear model:
data %>%
deconvolve_cellularity(
	!!.sample, !!.transcript, !!.abundance,
	method=method,
	reference = reference,
	action="get",
	...
)  %>%
	[..] %>%
	betareg::betareg(.my_formula, .)

Underlying method for the cox regression:
data %>%
deconvolve_cellularity(
	!!.sample, !!.transcript, !!.abundance,
	method=method,
	reference = reference,
	action="get",
	...
)  %>%
	[..] %>%
	mutate(.proportion_0_corrected = .proportion_0_corrected  %>% boot::logit()) %>%
	survival::coxph(.my_formula, .)
}
\examples{


	test_differential_cellularity(
	 tidybulk::counts_mini,
	    ~ condition,
	    sample,
	    transcript,
	    count,
	    cores = 1
	)



}
