% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_SE.R
\docType{methods}
\name{test_gene_enrichment}
\alias{test_gene_enrichment}
\alias{test_gene_enrichment,spec_tbl_df-method}
\alias{test_gene_enrichment,tbl_df-method}
\alias{test_gene_enrichment,tidybulk-method}
\alias{test_gene_enrichment,SummarizedExperiment-method}
\alias{test_gene_enrichment,RangedSummarizedExperiment-method}
\title{analyse gene enrichment with EGSEA}
\usage{
test_gene_enrichment(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  method = c("camera", "roast", "safe", "gage", "padog", "globaltest", "ora"),
  gene_collections = "all",
  species,
  cores = 10
)

\S4method{test_gene_enrichment}{spec_tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  method = c("camera", "roast", "safe", "gage", "padog", "globaltest", "ora"),
  gene_collections = "all",
  species,
  cores = 10
)

\S4method{test_gene_enrichment}{tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  method = c("camera", "roast", "safe", "gage", "padog", "globaltest", "ora"),
  gene_collections = "all",
  species,
  cores = 10
)

\S4method{test_gene_enrichment}{tidybulk}(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  method = c("camera", "roast", "safe", "gage", "padog", "globaltest", "ora"),
  gene_collections = "all",
  species,
  cores = 10
)

\S4method{test_gene_enrichment}{SummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  method = c("camera", "roast", "safe", "gage", "padog", "globaltest", "ora"),
  species,
  cores = 10
)

\S4method{test_gene_enrichment}{RangedSummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  method = c("camera", "roast", "safe", "gage", "padog", "globaltest", "ora"),
  species,
  cores = 10
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model}

\item{.sample}{The name of the sample column}

\item{.entrez}{The ENTREZ ID of the transcripts/genes}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.contrasts}{= NULL,}

\item{method}{A character vector. One or 3 or more methods to use in the testing (currently EGSEA errors if 2 are used). Type EGSEA::egsea.base() to see the supported GSE methods.}

\item{gene_collections}{A character vector. Used to determine which gene set collections to include in EGSEA buildIdx. It can take one or more of the following quoted values. MSigDB sets: "h", "c1", "c2", "c3", "c4", "c5", "c6","c7". c1" is human specific. KEGG sets: "Disease", "Metabolism", "Signaling". Default is "all", all MSigDB and KEGG gene set collections are used.}

\item{species}{A character. For example, human or mouse}

\item{cores}{An integer. The number of cores available}
}
\value{
A `tbl` object

A `tbl` object

A `tbl` object

A `tbl` object

A `tbl` object

A `tbl` object
}
\description{
test_gene_enrichment() takes as input a `tbl` formatted as | <SAMPLE> | <ENSEMBL_ID> | <COUNT> | <...> | and returns a `tbl` of gene set information
}
\details{
\lifecycle{maturing}

This wrapper executes ensemble gene enrichment analyses of the dataset using EGSEA (DOI:0.12688/f1000research.12544.1)


dge =
	data %>%
	keep_abundant(
		factor_of_interest = !!as.symbol(parse_formula(.formula)[[1]]),
		!!.sample, !!.entrez, !!.abundance
	) %>%

	# Make sure transcript names are adjacent
	[...] %>%
	as_matrix(rownames = !!.entrez) %>%
	edgeR::DGEList(counts = .)

idx =  buildIdx(entrezIDs = rownames(dge), species = species, msigdb.gsets = msigdb.gsets, 
               kegg.exclude = kegg.exclude)

dge %>%

	# Calculate weights
	limma::voom(design, plot = FALSE) %>%

	# Execute EGSEA
	egsea(
		contrasts = my_contrasts,
		baseGSEAs = method,
		gs.annots = idx,
		sort.by = "med.rank",
		num.threads = cores,
		report = FALSE
	)
}
\examples{
\dontrun{

df_entrez = symbol_to_entrez(tidybulk::counts_mini, .transcript = transcript, .sample = sample)
df_entrez = aggregate_duplicates(df_entrez, aggregation_function = sum, .sample = sample, .transcript = entrez, .abundance = count)

library("EGSEA")

	test_gene_enrichment(
		df_entrez,
		~ condition,
		.sample = sample,
		.entrez = entrez,
		.abundance = count,
         method = c("roast" , "safe", "gage"  ,  "padog" , "globaltest", "ora" ),
         gene_collections = "all",
		species="human",
		cores = 2
	)

}

}
