% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{test_gene_enrichment}
\alias{test_gene_enrichment}
\alias{test_gene_enrichment,spec_tbl_df-method}
\alias{test_gene_enrichment,tbl_df-method}
\alias{test_gene_enrichment,tidybulk-method}
\title{analyse gene enrichment with EGSEA}
\usage{
test_gene_enrichment(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  species,
  cores = 10
)

\S4method{test_gene_enrichment}{spec_tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  species,
  cores = 10
)

\S4method{test_gene_enrichment}{tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  species,
  cores = 10
)

\S4method{test_gene_enrichment}{tidybulk}(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  species,
  cores = 10
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model}

\item{.sample}{The name of the sample column}

\item{.entrez}{The ENTREZ ID of the transcripts/genes}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.contrasts}{= NULL,}

\item{species}{A character. For example, human or mouse}

\item{cores}{An integer. The number of cores available}
}
\value{
A `tbl` object

A `tbl` object

A `tbl` object

A `tbl` object
}
\description{
test_gene_enrichment() takes as input a `tbl` formatted as | <SAMPLE> | <ENSEMBL_ID> | <COUNT> | <...> | and returns a `tbl` with the additional transcript symbol column
}
\details{
\lifecycle{maturing}

This wrapper execute ensemble gene enrichment analyses of the dataset using EGSEA (DOI:0.12688/f1000research.12544.1)


dge =
	data %>%
	keep_abundant(		
		factor_of_interest = !!as.symbol(parse_formula(.formula)[[1]]),
		!!.sample, !!.entrez, !!.abundance 
	) %>%
	
	# Make sure transcript names are adjacent
	[...] %>%
	as_matrix(rownames = !!.entrez) %>%
	edgeR::DGEList(counts = .)

idx =  buildIdx(entrezIDs = rownames(dge), species = species)

dge %>%
	
	# Calculate weights
	limma::voom(design, plot = FALSE) %>%
	
	# Execute EGSEA
	egsea(
		contrasts = my_contrasts,
		baseGSEAs = egsea.base()[-c(6, 7, 8, 9, 12)],
		sort.by = "med.rank",
		num.threads = cores,
		report = FALSE
	)
}
\examples{
\dontrun{

df_entrez = symbol_to_entrez(tidybulk::counts_mini, .transcript = transcript, .sample = sample)
df_entrez = aggregate_duplicates(df_entrez, aggregation_function = sum, .sample = sample, .transcript = entrez, .abundance = count)

library("EGSEA")

	test_gene_enrichment(
		df_entrez,
		~ condition,
		.sample = sample,
		.entrez = entrez,
		.abundance = count,
		species="human",
		cores = 1
	)

}

}
