% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_gene_enrichment_bulk_EGSEA}
\alias{test_gene_enrichment_bulk_EGSEA}
\title{Get gene enrichment analyses using EGSEA}
\usage{
test_gene_enrichment_bulk_EGSEA(
  .data,
  .formula,
  .sample = NULL,
  .entrez,
  .abundance = NULL,
  .contrasts = NULL,
  method,
  gene_collections,
  species,
  cores = 10
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model}

\item{.sample}{The name of the sample column}

\item{.entrez}{The ENTREZ code of the transcripts/genes}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.contrasts}{A character vector. See edgeR makeContrasts specification for the parameter `contrasts`. If contrasts are not present the first covariate is the one the model is tested against (e.g., ~ factor_of_interest)}

\item{method}{A character vector. One or 3 or more methods to use in the testing (currently EGSEA errors if 2 are used). Type EGSEA::egsea.base() to see the supported GSE methods.}

\item{gene_collections}{A character vector. Used to determine which gene set collections to include in EGSEA buildIdx. It can take one or more of the following quoted values. MSigDB sets: "h", "c1", "c2", "c3", "c4", "c5", "c6","c7". c1" is human specific. KEGG sets: "Disease", "Metabolism", "Signaling". Default is "all", all MSigDB and KEGG gene set collections are used.}

\item{species}{A character. For example, human or mouse}

\item{cores}{An integer. The number of cores available}
}
\value{
A tibble with edgeR results
}
\description{
Get gene enrichment analyses using EGSEA
}
\keyword{internal}
