% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{test_gene_overrepresentation}
\alias{test_gene_overrepresentation}
\alias{test_gene_overrepresentation,spec_tbl_df-method}
\alias{test_gene_overrepresentation,tbl_df-method}
\alias{test_gene_overrepresentation,tidybulk-method}
\title{analyse gene over-representation with GSEA}
\usage{
test_gene_overrepresentation(.data, .sample = NULL, .entrez, .do_test, species)

\S4method{test_gene_overrepresentation}{spec_tbl_df}(.data, .sample = NULL, .entrez, .do_test, species)

\S4method{test_gene_overrepresentation}{tbl_df}(.data, .sample = NULL, .entrez, .do_test, species)

\S4method{test_gene_overrepresentation}{tidybulk}(.data, .sample = NULL, .entrez, .do_test, species)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.entrez}{The ENTREZ ID of the transcripts/genes}

\item{.do_test}{A boolean column name symbol. It indicates the transcript to check}

\item{species}{A character. For example, human or mouse. MSigDB uses the latin species names (e.g., \"Mus musculus\", \"Homo sapiens\")}
}
\value{
A `tbl` object

A `tbl` object

A `tbl` object

A `tbl` object
}
\description{
test_gene_overrepresentation() takes as input a `tbl` formatted as | <SAMPLE> | <ENSEMBL_ID> | <COUNT> | <...> | and returns a `tbl` with the GSEA statistics
}
\details{
\lifecycle{maturing}

This wrapper execute gene enrichment analyses of the dataset using a list of transcripts and GSEA. This wrapper uses clusterProfiler on the backend.
}
\examples{

df_entrez = symbol_to_entrez(tidybulk::counts_mini, .transcript = transcript, .sample = sample)
df_entrez = aggregate_duplicates(df_entrez, aggregation_function = sum, .sample = sample, .transcript = entrez, .abundance = count)
df_entrez = mutate(df_entrez, do_test = transcript \%in\% c("TNFRSF4", "PLCH2", "PADI4", "PAX7"))

	test_gene_overrepresentation(
		df_entrez,
		.sample = sample,
		.entrez = entrez,
		.do_test = do_test,
		species="Homo sapiens"
	)


}
