% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_SE.R
\docType{methods}
\name{test_stratification_cellularity}
\alias{test_stratification_cellularity}
\alias{test_stratification_cellularity,spec_tbl_df-method}
\alias{test_stratification_cellularity,tbl_df-method}
\alias{test_stratification_cellularity,tidybulk-method}
\alias{test_stratification_cellularity,SummarizedExperiment-method}
\alias{test_stratification_cellularity,RangedSummarizedExperiment-method}
\title{Test of stratification of biological replicates based on tissue composition, 
one cell-type at the time, using Kaplan-meier curves.}
\usage{
test_stratification_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  ...
)

\S4method{test_stratification_cellularity}{spec_tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  ...
)

\S4method{test_stratification_cellularity}{tbl_df}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  ...
)

\S4method{test_stratification_cellularity}{tidybulk}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  ...
)

\S4method{test_stratification_cellularity}{SummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  ...
)

\S4method{test_stratification_cellularity}{RangedSummarizedExperiment}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  ...
)
}
\arguments{
\item{.data}{A `tbl` (with at least three columns for sample, feature and 
transcript abundance) or `SummarizedExperiment` (more convenient if 
abstracted to tibble with library(tidySummarizedExperiment))}

\item{.formula}{A formula representing the desired linear model. The formula 
can be of two forms: multivariable (recommended) or univariable 
Respectively: \"factor_of_interest ~ .\" or \". ~ factor_of_interest\". 
The dot represents cell-type proportions, and it is mandatory. If censored 
regression is desired (coxph) the formula should be of the form \"survival::Surv\(y, dead\) ~ .\"}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{method}{A string character. Either \"cibersort\", \"epic\" or \"llsr\". 
The regression method will be chosen based on being multivariable: lm or 
cox-regression (both on logit-transformed proportions); or univariable: beta 
or cox-regression (on logit-transformed proportions). See .formula for 
multi- or univariable choice.}

\item{reference}{A data frame. The transcript/cell_type data frame of integer 
transcript abundance}

\item{...}{Further parameters passed to the method deconvolve_cellularity}
}
\value{
A consistent object (to the input) with additional columns for the 
statistics from the hypothesis test (e.g.,  log fold change, p-value and 
false discovery rate).

A consistent object (to the input) with additional columns for the 
statistics from the hypothesis test (e.g.,  log fold change, p-value and 
false discovery rate).

A consistent object (to the input) with additional columns for the 
statistics from the hypothesis test (e.g.,  log fold change, p-value and 
false discovery rate).
}
\description{
test_stratification_cellularity() takes as input A `tbl` 
(with at least three columns for sample, feature and transcript abundance) 
or `SummarizedExperiment` (more convenient if abstracted to tibble with 
library(tidySummarizedExperiment)) and returns a consistent object 
(to the input) with additional columns for the statistics from the hypothesis test.
}
\details{
`r lifecycle::badge("maturing")`

This routine applies a deconvolution method (e.g., Cibersort; DOI: 10.1038/nmeth.3337)
and passes the proportions inferred into a generalised linear model 
(DOI:dx.doi.org/10.1007/s11749-010-0189-z)
or a cox regression model (ISBN: 978-1-4757-3294-8)


Underlying method for the test:
data |>
deconvolve_cellularity(
	!!.sample, !!.transcript, !!.abundance,
	method=method,
	reference = reference,
	action="get",
	...
)  %>%
	[..] |>
	mutate(.high_cellularity = .proportion > median(.proportion)) |>
	survival::survdiff(data = data, .my_formula)
}
\examples{
tidybulk::se_mini |>
test_stratification_cellularity(
	survival::Surv(days, dead) ~ .,
	cores = 1
)



}
