% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{tidyBulk}
\alias{tidyBulk}
\title{Creates a `tt` object from a `tbl``}
\usage{
tidyBulk(.data, .sample, .transcript, .abundance, .abundance_scaled = NULL)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.abundance_scaled}{The name of the transcript/gene scaled abundance column}
}
\value{
A `tidyBulk` object
}
\description{
tidyBulk() creates a `tt` object from a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |
}
\details{
\lifecycle{maturing}

This function created a tidyBulk object and is useful if you want
to avoid to specify .sample, .transcript and .abundance arguments all the times.
The tidyBulk object have an attribute called tt_columns where these three
arguments are stored as metadata. They can be extracted as attr(<object>, "tt_columns").
}
\examples{




my_tt =  tidyBulk(tidyBulk::counts_mini, sample, transcript, count)


}
