% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidyBulk,SummarizedExperiment-method}
\alias{tidyBulk,SummarizedExperiment-method}
\title{tidyBulk}
\usage{
\S4method{tidyBulk}{SummarizedExperiment}(.data, .sample, .transcript, .abundance, .abundance_scaled = NULL)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.abundance_scaled}{The name of the transcript/gene scaled abundance column}
}
\value{
A `tidyBulk` object
}
\description{
tidyBulk
}
