% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_cibersort}
\alias{X_cibersort}
\title{Cibersort reference}
\format{An object of class \code{data.frame} with 547 rows and 22 columns.}
\usage{
X_cibersort
}
\description{
Cibersort reference
}
\keyword{datasets}
